% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{aggregate.sf}
\alias{aggregate.sf}
\title{aggregate sf objects over a spatial region}
\usage{
\method{aggregate}{sf}(x, by, FUN, ..., extensive)
}
\arguments{
\item{x}{object of class \code{sf}, for which we want to aggregate attributes}

\item{by}{object of class \code{sf} or \code{sfc}, with the target geometries}

\item{FUN}{aggregation function, should not be set}

\item{...}{arguments, passed on to FUN}

\item{extensive}{logical; if TRUE, the attribute variables are assumed to be spatially extensive.}
}
\description{
aggregate sf objects over a spatial region
}
\examples{
nc = st_read(system.file("shape/nc.shp", package="sf"))
g = sf:::st_makegrid(nc, n = c(20,10))
a1 = aggregate(nc["BIR74"], g)
sum(a1$BIR74) / sum(nc$BIR74) # not close to one: property is assumed spatially intensive
a2 = aggregate(nc["BIR74"], g, extensive = TRUE)
sum(a2$BIR74) / sum(nc$BIR74)
a1$intensive = a1$BIR74
a1$extensive = a2$BIR74
plot(a1[c("intensive", "extensive")])
}

