% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{as.sfc}
\alias{as.sfc}
\title{convert foreign geometry object to an sfc object}
\usage{
as.sfc(x, ...)
}
\arguments{
\item{x}{object to convert}

\item{...}{further arguments}
}
\description{
convert foreign geometry object to an sfc object
}

