% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{as.sfc.SpatialPolygons}
\alias{as.sfc.SpatialPolygons}
\title{coerce SpatialPolygons object into sfc}
\usage{
\method{as.sfc}{SpatialPolygons}(x, ..., forceMulti = FALSE)
}
\arguments{
\item{x}{object of class \link[sp]{SpatialPolygons}}

\item{...}{ignored}

\item{forceMulti}{logical: force coercion into MultiPolygon objects (TRUE), or autodetect}
}
\description{
coerce SpatialPolygons object into sfc
}

