% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{as.WKT}
\alias{as.WKT}
\title{Return Well-known Text representation of simple feature item}
\usage{
as.WKT(x)
}
\arguments{
\item{x}{object of class sfi}
}
\description{
Return Well-known Text representation of simple feature item
}

