% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{as_Spatial}
\alias{as_Spatial}
\title{Convert to sp object}
\usage{
as_Spatial(from, cast = TRUE)
}
\arguments{
\item{from}{sfc to convert}

\item{cast}{logical. Coerce GEOMETRIES to a type using `st_cast` (default TRUE)}
}
\description{
Convert to sp object
}
