% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox}
\alias{bbox}
\title{Return bounding of a simple feature or simple feature set}
\usage{
bbox(obj)
}
\arguments{
\item{obj}{object to compute the bounding box from}
}
\description{
Return bounding of a simple feature or simple feature set
}

