% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{bind}
\alias{bind}
\alias{cbind.sf}
\alias{rbind.sf}
\title{bind rows (features) of sf objects}
\usage{
rbind.sf(..., deparse.level = 1)

cbind.sf(..., deparse.level = 1)
}
\arguments{
\item{...}{objects to bind}

\item{deparse.level}{integer; see \link[base]{rbind}}

\item{...}{objects to bind}

\item{deparse.level}{integer; see \link[base]{cbind}}
}
\value{
if \code{cbind} is called with multiple \code{sf} objects, it warns and removes all but the first geometry column from the input objects.
}
\description{
bind rows (features) of sf objects

bind columns (variables) of sf objects
}

