% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{methods}
\name{dbDataType,PqConnection,sf-method}
\alias{dbDataType,PqConnection,sf-method}
\title{Determine database type for R vector.}
\usage{
\S4method{dbDataType}{PqConnection,sf}(dbObj, obj)
}
\arguments{
\item{dbObj}{Postgres driver or connection.}

\item{obj}{Object to convert}
}
\description{
Determine database type for R vector.
}
