% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{dplyr}
\alias{arrange_.sf}
\alias{distinct_.sf}
\alias{do_.sf}
\alias{dplyr}
\alias{filter_.sf}
\alias{group_by_.sf}
\alias{mutate_.sf}
\alias{rename_.sf}
\alias{select_.sf}
\alias{slice_.sf}
\alias{transmute_.sf}
\title{dplyr verb methods for sf objects}
\usage{
filter_.sf(.data, ..., .dots)

arrange_.sf(.data, ..., .dots)

distinct_.sf(.data, ..., .dots, .keep_all = FALSE)

do_.sf(.data, ..., .dots, .chunk_size = 10000L)

group_by_.sf(.data, ..., .dots, add = FALSE)

mutate_.sf(.data, ..., .dots)

transmute_.sf(.data, ..., .dots)

select_.sf(.data, ..., .dots)

rename_.sf(.data, ..., .dots)

slice_.sf(.data, ..., .dots)
}
\arguments{
\item{.data}{data object of class \link{sf}}

\item{...}{other arguments}

\item{.dots}{see corresponding function in package \code{dplyr}}

\item{.keep_all}{see corresponding function in dplyr}

\item{.chunk_size}{see corresponding function in dplyr}

\item{add}{see corresponding function in dplyr}
}
\description{
dplyr verb methods for sf objects
}

