% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{data}
\name{extension_map}
\alias{extension_map}
\title{Map extension to driver}
\format{An object of class \code{list} of length 23.}
\usage{
extension_map
}
\description{
Map extension to driver
}
\keyword{datasets}
