% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{find_database_srid}
\alias{find_database_srid}
\title{Find srid in a database by using the srid}
\usage{
find_database_srid(conn, crs_local = st_crs(srid), srid = epsg(crs_local))
}
\arguments{
\item{conn}{Dababase connection (e.g. `DBI`)}

\item{srid}{An integer descriing the srid to fetch}
}
\value{
a `crs`
}
\description{
Find srid in a database by using the srid
}
