% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{find_database_srtext}
\alias{find_database_srtext}
\title{Find database projection using srtext (wkt)}
\usage{
find_database_srtext(
  conn,
  crs_local = st_crs(wkt),
  wkt = st_as_text(crs_local)
)
}
\description{
Find database projection using srtext (wkt)
}
