% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{gdal_compressors}
\alias{gdal_compressors}
\title{List GDAL compressors and decompressors}
\usage{
gdal_compressors()
}
\value{
named list with two character vectors, containing compressors and decompressors
}
\description{
List GDAL compressors and decompressors
}
