% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{gdal_inv_geotransform}
\alias{gdal_inv_geotransform}
\title{retrieve the inverse of a gdal geotransform}
\usage{
gdal_inv_geotransform(gt)
}
\arguments{
\item{gt}{double vector of length 6}
}
\description{
retrieve the inverse of a gdal geotransform
}
