% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{gdal_metadata}
\alias{gdal_metadata}
\alias{gdal_metadata}
\alias{gdal_subdatasets}
\title{get metadata of a raster layer}
\usage{
gdal_metadata(file, domain_item = character(0), options = character(0),
  parse = TRUE)

gdal_subdatasets(file, options = character(0), name = TRUE)
}
\arguments{
\item{file}{file name}

\item{domain_item}{character vector of length 0, 1 (with domain), or 2 (with domain and item); use \code{""} for the default domain, use \code{NA_character_} to query the domain names.}

\item{options}{character; character vector with data open options}

\item{parse}{logical; should metadata be parsed into a named list (\code{TRUE}) or returned as character data?}

\item{name}{logical; retrieve name of subdataset? If \code{FALSE}, retrieve description}
}
\value{
named list with metadata items

\code{gdal_subdatasets} returns a zero-length list if \code{file} does not have subdatasets, and else a named list with subdatasets.
}
\description{
get metadata of a raster layer
}
\examples{
#f = system.file("tif/L7_ETMs.tif", package="stars")
f = system.file("nc/avhrr-only-v2.19810901.nc", package = "stars")
gdal_metadata(f)
gdal_metadata(f, NA_character_)
# try(gdal_metadata(f, "wrongDomain"))
# gdal_metadata(f, c("", "AREA_OR_POINT"))
}
