% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{gdal_read}
\alias{gdal_read}
\title{read gdal raster file (not to be called by users, but to be used by st_stars)}
\usage{
gdal_read(x, options = character(0), driver = character(0))
}
\arguments{
\item{x}{character vector, possibly of length larger than 1 when more than one raster is read}

\item{options}{character}

\item{driver}{character; when empty vector, driver is auto-detected.}
}
\description{
read gdal raster file (not to be called by users, but to be used by st_stars)
}
