% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\docType{methods}
\name{geometry,sf-method}
\alias{geometry,sf-method}
\title{get geometry from sf object}
\usage{
\S4method{geometry}{sf}(obj)
}
\arguments{
\item{obj}{object of class \code{sf}}
}
\description{
get geometry from sf object
}

