% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos.R
\name{geos}
\alias{geos}
\alias{st_distance}
\alias{st_geos_binop}
\alias{st_is_valid}
\title{Compute distance between pairs of simple feature geometries}
\usage{
st_distance(sf1, sf2 = sf1)

st_is_valid(sf)

st_geos_binop(op = "intersects", sf1, sf2 = sf1, par = 0, sparse = TRUE)
}
\arguments{
\item{sf1}{first simple feature (sf) or simple feature geometry (sfc) collection}

\item{sf2}{second simple feature (sf) or simple feature geometry (sfc) collection}

\item{sf}{simple feautures (sf) or simple feature geometries (sfc) object}

\item{op}{character; one of "intersects", "disjoint", "touches", "crosses", "within", "contains", "overlaps", "equals", "covers", "coveredBy", "equalsExact", or "isWithinDistance", "relate", "distance"}

\item{par}{numeric; parameter used for "equalsExact" (margin) and "isWithinDistance"}

\item{sparse}{logical; should a sparse matrix be returned (TRUE) or a dense matrix?}
}
\value{
numeric matrix of dimension length(sf1) x length(sf2)

logical vector

matrix (sparse or dense) of type \code{character} for op \code{relate}, \code{numeric} for \code{distance}, and \code{logical} for all others; matrix has dimension \code{sf1} x \code{sf2}.
}
\description{
Compute distance between pairs of simple feature geometries
}

