% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{is_driver_available}
\alias{is_driver_available}
\title{Check if driver is available}
\usage{
is_driver_available(drv, drivers = st_drivers())
}
\arguments{
\item{drv}{character. Name of driver}

\item{drivers}{data.frame. Table containing driver names and support. Default
is from \code{\link{st_drivers}}}
}
\description{
Search through the driver table if driver is listed
}

