% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{is_driver_can}
\alias{is_driver_can}
\title{Check if a driver can perform an action}
\usage{
is_driver_can(drv, drivers = st_drivers(), operation = "write")
}
\arguments{
\item{drv}{character. Name of driver}

\item{drivers}{data.frame. Table containing driver names and support. Default
is from \code{\link{st_drivers}}}

\item{operation}{character. What action to check}
}
\description{
Search through the driver table to match a driver name with 
an action (e.g. \code{"write"}) and check if the action is supported.
}
