% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{is_geometry_column}
\alias{is_geometry_column}
\title{Check if the columns could be of a coercable type for sf}
\usage{
is_geometry_column(con, x, classes = "")
}
\arguments{
\item{con}{database connection}

\item{x}{inherits data.frame}

\item{classes}{classes inherited}
}
\description{
Check if the columns could be of a coercable type for sf
}
