% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{is_parallel_available}
\alias{is_parallel_available}
\title{Is parallel processing available?}
\usage{
is_parallel_available()
}
\value{
logical; indicating if parallel processing is available
}
\description{
Is parallel processing available?
}
\examples{
is_parallel_available()
}
