% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{LINESTRING}
\alias{LINESTRING}
\title{Create a linestring simple feature from a numeric matrix}
\usage{
LINESTRING(x, third = "Z", ...)
}
\arguments{
\item{x}{numeric matrix with 2, 3 or 4 columns}

\item{third}{character, indicating what the third dimension refers to ("Z" or "M")}

\item{...}{ignored}
}
\description{
Create a linestring simple feature from a numeric matrix
}

