% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_aware.R
\name{make_sf_aware}
\alias{make_sf_aware}
\title{for packages: use the sticky geometry [ behaviour of sf in package code}
\usage{
make_sf_aware(env = parent.frame(), value = TRUE)
}
\description{
for packages: use the sticky geometry [ behaviour of sf in package code
param env environment
value logical; default \code{TRUE}
}
