% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{merge.sf}
\alias{merge.sf}
\title{merge method for sf and data.frame object}
\usage{
\method{merge}{sf}(x, y, ...)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{y}{object of class \code{data.frame}}

\item{...}{arguments passed on to \code{merge.data.frame}}
}
\description{
merge method for sf and data.frame object
}
\examples{
a = data.frame(a = 1:3, b = 5:7)
st_geometry(a) = st_sfc(st_point(c(0,0)), st_point(c(1,1)), st_point(c(2,2)))
b = data.frame(x = c("a", "b", "c"), b = c(2,5,6))
merge(a, b)
merge(a, b, all = TRUE)
}

