% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{MULTIPOLYGON}
\alias{MULTIPOLYGON}
\title{Create a multipolygon simple feature from a list of lists with coordinate matrices}
\usage{
MULTIPOLYGON(x, third = "Z", ...)
}
\arguments{
\item{x}{list with numeric matrices with 2, 3 or 4 columns}

\item{third}{character, indicating what the third dimension refers to ("Z" or "M")}

\item{...}{ignored}
}
\description{
Create a multipolygon simple feature from a list of lists with coordinate matrices
}

