% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{names<-.sf}
\alias{names<-.sf}
\title{Rename sf object}
\usage{
\method{names}{sf}(x) <- value
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\description{
Rename sf object
}
