% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{number_processors}
\alias{number_processors}
\title{Number of processors}
\usage{
number_processors()
}
\value{
integer; number of logical processors
}
\description{
Number of processors
}
\examples{
number_processors()
}
