% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{openmp_version}
\alias{openmp_version}
\title{OpenMP version}
\usage{
openmp_version()
}
\value{
character; version of OpenMP library available
}
\description{
OpenMP version
}
\examples{
openmp_version()
}
