% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sf}
\alias{plot.sf}
\title{plot sf object}
\usage{
\method{plot}{sf}(x, y, ..., xlim, ylim, xaxs, yaxs, lab, asp = NA,
  axes = TRUE, add = FALSE)
}
\arguments{
\item{x}{object of class sf}

\item{y}{ignored}

\item{...}{further specifications, see \link{plot}}

\item{xlim}{see \link{plot.window}}

\item{ylim}{see \link{plot.window}}

\item{xaxs}{see \link{plot.window}}

\item{yaxs}{see \link{plot.window}}

\item{lab}{see \link{plot.window}}

\item{asp}{aspect ratio, see \link{plot.window}}

\item{axes}{logical; plot axes?}

\item{add}{logical; add to existing plot?}
}
\description{
plot sf object
}

