% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{POINT}
\alias{POINT}
\title{Create a point simple feature from a numeric vector}
\usage{
POINT(x, third = "Z", ...)
}
\arguments{
\item{x}{numeric vector of length 2, 3 or 4}

\item{third}{character, indicating what a 3-dimensional point refers to ("Z" or "M")}

\item{...}{ignored}
}
\description{
Create a point simple feature from a numeric vector
}

