% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{POLYGON}
\alias{POLYGON}
\title{Create a polygon simple feature from a list with coordinate matrices}
\usage{
POLYGON(x, third = "Z", ...)
}
\arguments{
\item{x}{list with numeric matrices with 2, 3 or 4 columns}

\item{third}{character, indicating what the third dimension refers to ("Z" or "M")}

\item{...}{ignored}
}
\description{
Create a polygon simple feature from a list with coordinate matrices
}

