% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{data}
\name{prefix_map}
\alias{prefix_map}
\title{Map prefix to driver}
\format{An object of class \code{list} of length 11.}
\usage{
prefix_map
}
\description{
Map prefix to driver
}
\note{
The \code{sf} driver is a special prefix that reads data from
the \code{sf} examples. See \code{st_example()} for details.
}
\keyword{datasets}
