% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{proj_tools}
\alias{proj_tools}
\alias{sf_proj_search_paths}
\alias{sf_proj_network}
\title{Manage PROJ settings}
\usage{
sf_proj_search_paths(paths = character(0))

sf_proj_network(enable = FALSE, url = character(0))
}
\arguments{
\item{paths}{the search path to be set; omit if no paths need to be set}

\item{enable}{logical; set this to enable (TRUE) or disable (FALSE) the proj network search facility}

\item{url}{character; use this to specify and override the default proj network CDN}
}
\value{
`sf_proj_search_paths()` returns the search path (possibly after setting to paths)

`sf_proj_network` when called without arguments returns a logical indicating whether
network search of datum grids is enabled, when called with arguments it returns a character
vector with the URL of the CDN used (or specified with `url`).
}
\description{
Manage PROJ search path and network settings
}
