% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R
\name{rawToHex}
\alias{rawToHex}
\title{Convert raw vector(s) into hexadecimal character string(s)}
\usage{
rawToHex(x)
}
\arguments{
\item{x}{raw vector, or list with raw vectors}
}
\description{
Convert raw vector(s) into hexadecimal character string(s)
}
