% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelOptions.R
\name{set_mc_option}
\alias{set_mc_option}
\title{Options for parallel support}
\usage{
set_mc_option(value)
}
\arguments{
\item{value}{logical; indicates the availability of multicore clusters, set to TRUE on package load unless platform is Windows, in which case FALSE}
}
\description{
Provides support for the use of parallel computation from the parallel package
}
\details{
Parallel processing options in the sf package are held in an environment local to the package namespace and not exported. Option values are set and retrieved with pairs of access functions, get and set. The \code{mc} option is set by default to FALSE on Windows systems, as they cannot fork the R session; by default it is TRUE on other systems, but may be set FALSE. If \code{mc} is FALSE, the \code{Cluster} option is used: if \code{mc} is FALSE and the \code{Cluster} option is NULL no parallel computing is done, or the \code{Cluster} option is passed a \dQuote{cluster} object created by the parallel or snow package for access without being passed as an argument. The \code{cores} option is set to NULL by default, and can be used to store the number of cores to use as an integer. If \code{cores} is NULL, facilities from the parallel package will not be used.
}
