% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_gpkg.R
\name{set_timestamp_gpkg}
\alias{set_timestamp_gpkg}
\title{Set explicit timestamp in a GeoPackage 'gpkg_contents' table}
\usage{
set_timestamp_gpkg(dsn, timestamp = Sys.time(), verbose = TRUE)
}
\arguments{
\item{dsn}{the path to the GeoPackage file (*.gpkg)}

\item{timestamp}{a \code{Date} or \code{POSIXct} object, used to generate
the timestamp.
For a \code{Date} object, time will be set at 00:00:00 UTC.
Defaults to system time, however must be set explicitly for reproducible
workflows.}

\item{verbose}{Logical.
Prints a message with the number of affected rows in \code{gpkg_contents}.}
}
\value{
Value \code{0} is returned invisibly.
}
\description{
Overwrites all timestamps (column \code{last_change}) of the
\code{gpkg_contents} table in an existing GeoPackage file.
As such the function assists in making a binary-reproducible GeoPackage file.
}
\details{
Internally the timestamp is converted to a specific ISO 8601 format
that is required by the GeoPackage standard.
}
\author{
Floris Vanderhaeghe, \url{https://github.com/florisvdh}
}
