% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setprojpath.R
\name{setprojpath}
\alias{setprojpath}
\title{Set the PROJ_LIB path to the 'proj' subfolder of `sf` installation
folder on windows () and the GDAL_DATA folder to the 'GDAL_DATA' folder
of `sf` and autoatically restore the previous (system) value upon
exiting from the caller function. This avoids the need of modifying
values of the two variables on load.}
\usage{
setprojpath()
}
\description{
Set the PROJ_LIB path to the 'proj' subfolder of `sf` installation
folder on windows () and the GDAL_DATA folder to the 'GDAL_DATA' folder
of `sf` and autoatically restore the previous (system) value upon
exiting from the caller function. This avoids the need of modifying
values of the two variables on load.
}
