% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{sf-defunct}
\alias{sf-defunct}
\alias{st_read_db}
\alias{st_write_db}
\title{Deprecated functions in \code{sf}}
\usage{
st_read_db(
  conn = NULL,
  table = NULL,
  query = NULL,
  geom_column = NULL,
  EWKB = TRUE,
  ...
)

st_write_db(
  conn = NULL,
  obj,
  table = deparse(substitute(obj)),
  ...,
  drop = FALSE,
  append = FALSE
)
}
\arguments{
\item{conn}{open database connection}

\item{table}{table name}

\item{query}{SQL query to select records; see details}

\item{geom_column}{deprecated. Geometry column name}

\item{EWKB}{logical; is the WKB of type EWKB? if missing, defaults to
\code{TRUE}}

\item{...}{parameter(s) passed on to \link{st_as_sf}}
}
\description{
These functions are provided for compatibility with older version of \code{sf}.
They will eventually be completely removed.
\itemize{
\item Use \code{\link[=st_read]{st_read()}} instead of \code{st_read_db()}.
\item Use \code{\link[=st_write]{st_write()}} instead_of \code{st_write_db()}
}
}
\details{
The \code{geom_column} argument is deprecated. The function will
automatically find the \code{geometry} type columns. For the \code{RPostgreSQL} drivers
it will try to cast all the character columns, which can be long for very wide
tables.
}
\keyword{internal}
