\name{sf-package}
\alias{sf-package}

\alias{sf}

\docType{package}

\title{The \code{sf} package}

\description{

Package \code{sf} provides classes and methods to work with spatial
vector data (points, lines, polygons).  The data structures and
methods it implements follow the "Simple Feature Access" standard,
which is also found in many other softwares like GDAL, GEOS, QGIS,
PostGIS and other spatial databases. For introductions into the
package please consult:

Pebesma, E. and Bivand, R. (2023). Spatial Data
Science: With Applications in R. Chapman and Hall/CRC.
https://doi.org/10.1201/9780429459016 which is also found freely
online at https://r-spatial.org/book/

or

Pebesma, E., 2018. Simple Features for R: Standardized Support
for Spatial Vector Data. The R Journal 10 (1), 439-446,
https://doi.org/10.32614/RJ-2018-009 (open access)

}
