% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_types.R
\docType{data}
\name{sf.types}
\alias{sf.types}
\title{data.frame listing all the simple feature types}
\format{An object of class \code{data.frame} with 72 rows and 5 columns.}
\usage{
sf.types
}
\description{
data.frame listing all simple feature types
}
\keyword{datasets}

