% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{sf_extSoftVersion}
\alias{sf_extSoftVersion}
\title{Provide the external dependencies versions of the libraries linked to sf}
\usage{
sf_extSoftVersion()
}
\description{
Provide the external dependencies versions of the libraries linked to sf
}
