% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_aware.R
\name{sf_make_aware}
\alias{sf_make_aware}
\alias{sf_make_aware_pkg}
\title{for packages: use the sticky geometry [ behaviour of sf in package code}
\usage{
sf_make_aware(env = parent.frame(), value = TRUE)

sf_make_aware_pkg(pkg)
}
\arguments{
\item{env}{environment}

\item{value}{logical; default \code{TRUE}}

\item{pkg}{package name}
}
\description{
for packages: use the sticky geometry [ behaviour of sf in package code
}
