% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sf_proj_search_paths}
\alias{sf_proj_search_paths}
\title{Get or set the PROJ search paths}
\usage{
sf_proj_search_paths(paths = NULL)
}
\arguments{
\item{paths}{the search path to be set; omit if no paths need to be set}
}
\value{
the search path (possibly after setting to paths)
}
\description{
Get or set the PROJ search paths
}
