% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sf_project}
\alias{sf_project}
\title{directly transform a set of coordinates}
\usage{
sf_project(from, to, pts, warn = TRUE)
}
\arguments{
\item{from}{character; proj4string of pts}

\item{to}{character; target coordinate reference system}

\item{pts}{two-column numeric matrix, or object that can be coerced into a matrix}

\item{warn}{logical; warn if invalid (NA) values are generated}
}
\value{
two-column numeric matrix with transformed/converted coordinates, returning invalid values as \code{NA}
}
\description{
directly transform a set of coordinates
}
