% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{sfc}
\alias{sfc}
\title{verify simple feature}
\usage{
sfc(lst, epsg = as.integer(NA), proj4string = as.character(NA))
}
\arguments{
\item{lst}{list with simple feature objects}

\item{epsg}{integer; epsg code}

\item{proj4string}{character; describing the coordinate reference systems in PROJ.4 syntax}
}
\description{
verifies simple feature list column's contents, and sets class
}
\examples{
pt1 = POINT(c(0,1))
pt2 = POINT(c(1,1))
(sfc = sfc(list(pt1, pt2)))
d = data.frame(a = 1:2)
}

