% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr.R
\docType{data}
\name{st_agr}
\alias{NA_agr_}
\alias{st_agr}
\alias{st_agr<-}
\alias{st_set_agr}
\title{get or set relation_to_geometry attribute of an \code{sf} object}
\format{An object of class \code{factor} of length 1.}
\usage{
NA_agr_

st_agr(x, ...)

st_agr(x) <- value

st_set_agr(x, value)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{...}{ignored}

\item{value}{character, or factor with appropriate levels; if named, names should correspond to the non-geometry list-column columns of \code{x}}
}
\description{
get or set relation_to_geometry attribute of an \code{sf} object
}
\details{
\code{NA_agr_} is the \code{agr} object with a missing value.
}
\keyword{datasets}

