% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{ST_as.sf.Spatial}
\alias{ST_as.sf.Spatial}
\title{convert foreign object to an sf object}
\usage{
\method{ST_as.sf}{Spatial}(x)
}
\arguments{
\item{x}{object deriving from \link[sp]{Spatial}}
}
\description{
convert foreign object to an sf object
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
summary(ST_as.sf(meuse))
summary(ST_as.sf(meuse.grid))
summary(ST_as.sf(meuse.area))
summary(ST_as.sf(meuse.riv))
summary(ST_as.sf(as(meuse.riv, "SpatialLines")))
pol.grd = as(meuse.grid, "SpatialPolygonsDataFrame")
summary(ST_as.sf(pol.grd))
summary(ST_as.sf(as(pol.grd, "SpatialLinesDataFrame")))
}

