% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{ST_as.sfc.SpatialLines}
\alias{ST_as.sfc.SpatialLines}
\title{coerce SpatialLines object into sfc}
\usage{
\method{ST_as.sfc}{SpatialLines}(x, ..., forceMulti = FALSE)
}
\arguments{
\item{x}{object of class \link[sp]{SpatialLines}}

\item{...}{ignored}

\item{forceMulti}{logical: force coercion into MultiPolygon objects (TRUE), or autodetect}
}
\description{
coerce SpatialLines object into sfc
}

