% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R
\name{ST_as.WKB}
\alias{ST_as.WKB}
\alias{ST_as.WKB.sfc}
\alias{ST_as.WKB.sfi}
\title{convert sfc object to an WKB object}
\usage{
ST_as.WKB(x, ...)

\method{ST_as.WKB}{sfc}(x, ..., endian = .Platform$endian)

\method{ST_as.WKB}{sfi}(x, ..., endian = .Platform$endian)
}
\arguments{
\item{x}{object to convert}

\item{...}{further arguments}

\item{endian}{character; either "big" or "little"; default: use that of platform}
}
\description{
convert sfc object to an WKB object
}

