% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{ST_as.WKT}
\alias{ST_as.WKT}
\title{Return Well-known Text representation of simple feature item}
\usage{
ST_as.WKT(x)
}
\arguments{
\item{x}{object of class sfi}
}
\description{
Return Well-known Text representation of simple feature item
}

