% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{st_as_grob}
\alias{st_as_grob}
\title{Convert sf* object to a grob}
\usage{
st_as_grob(x, ..., units = "native")
}
\arguments{
\item{x}{object to be converted into an object class \code{grob}}

\item{...}{passed on to the xxxGrob function, e.g. \code{gp = gpar(col = 'red')}}

\item{units}{units; see \link[grid]{unit}}
}
\description{
Convert sf* object to an grid graphics object (grob)
}
