% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R, R/wkt.R
\name{st_as_text}
\alias{st_as_text}
\alias{st_as_text.crs}
\alias{st_as_text}
\alias{st_as_text}
\alias{st_as_text.sfg}
\alias{st_as_text}
\alias{st_as_text.sfc}
\title{Return Well-known Text representation of simple feature geometry or coordinate reference system}
\usage{
\method{st_as_text}{crs}(x, ..., pretty = FALSE)

st_as_text(x, ...)

\method{st_as_text}{sfg}(x, ...)

\method{st_as_text}{sfc}(x, ..., EWKT = FALSE)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfc} or \code{crs}}

\item{...}{passed on to WKT_name}

\item{pretty}{logical; if TRUE, print human-readable well-known-text representation of a coordinate reference system}

\item{EWKT}{logical; if TRUE, print SRID=xxx; before the WKT string if \code{epsg} is available}
}
\description{
Return Well-known Text representation of simple feature geometry or coordinate reference system
}
\details{
To suppress printing of SRID, \code{EWKT=FALSE} can be passed as parameter.
}
\examples{
st_as_text(st_point(1:2))
}
