% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R
\name{st_as_wkb}
\alias{st_as_wkb}
\alias{st_as_wkb.sfc}
\alias{st_as_wkb.sfi}
\title{convert sfc object to an WKB object}
\usage{
st_as_wkb(x, ...)

\method{st_as_wkb}{sfc}(x, ..., endian = .Platform$endian)

\method{st_as_wkb}{sfi}(x, ..., endian = .Platform$endian)
}
\arguments{
\item{x}{object to convert}

\item{...}{further arguments}

\item{endian}{character; either "big" or "little"; default: use that of platform}
}
\description{
convert sfc object to an WKB object
}

