% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{st_as_wkt}
\alias{st_as_wkt}
\alias{st_as_wkt.sfc}
\alias{st_as_wkt.sfi}
\title{Return Well-known Text representation of simple feature item}
\usage{
st_as_wkt(x, ...)

\method{st_as_wkt}{sfi}(x, ...)

\method{st_as_wkt}{sfc}(x, ..., EWKT = FALSE)
}
\arguments{
\item{x}{object of class sfi or sfc}

\item{...}{passed on to WKT_name}

\item{EWKT}{logical; if TRUE, print SRID=xxx; before the WKT string if epsg is available}
}
\description{
Return Well-known Text representation of simple feature item
}
\details{
to suppress printing of SRID, \code{EWKT=FALSE} can be passed as parameter
}
\examples{
st_as_wkt(st_point(1:2))
}

