% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{st_bbox}
\alias{st_bbox}
\title{Return bounding of a simple feature or simple feature set}
\usage{
st_bbox(obj)
}
\arguments{
\item{obj}{object to compute the bounding box from}
}
\value{
a numeric vector of length four, with \code{xmin}, \code{ymin}, \code{xmax}
and \code{ymax} values; if \code{obj} is of class \code{sf} or \code{sfc}, the object
returned has a class \code{bbox}, an attribute \code{crs} and a method to print the
bbox and an \code{st_crs} method to retrieve the coordinate reference system 
corresponding to \code{obj} (and hence the bounding box).
}
\description{
Return bounding of a simple feature or simple feature set
}

