% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R, R/cast-individual.R
\name{st_cast}
\alias{st_cast}
\alias{st_cast.LINESTRING}
\alias{st_cast.MULTILINESTRING}
\alias{st_cast.MULTIPOINT}
\alias{st_cast.MULTIPOLYGON}
\alias{st_cast.POINT}
\alias{st_cast.POLYGON}
\title{Cast geometry to another type: either simplify, or cast explicitly}
\usage{
st_cast(x, to, ...)

\method{st_cast}{MULTIPOLYGON}(x, to)

\method{st_cast}{MULTILINESTRING}(x, to)

\method{st_cast}{MULTIPOINT}(x, to)

\method{st_cast}{POLYGON}(x, to)

\method{st_cast}{LINESTRING}(x, to)

\method{st_cast}{POINT}(x, to)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfc} or \code{sf}}

\item{to}{character; target type, if missing, simplification is tried; when \code{x} is of type \code{sfg} (i.e., a single geometry) then \code{to} needs to be specified.}

\item{...}{ignored}
}
\value{
object of class \code{to} if successful, or unmodified object if unsuccesful. If information gets lost while type casting, a warning is raised.
}
\description{
Cast geometry to another type: either simplify, or cast explicitly
}
\examples{
s = st_multipoint(rbind(c(1,0)))
st_cast(s, "POINT")
example(st_read)
mpl <- nc$geometry[[4]]
#st_cast(x) ## error 'argument "to" is missing, with no default'
cast_all <- function(xg) {
  lapply(c("MULTIPOLYGON", "MULTILINESTRING", "MULTIPOINT", "POLYGON", "LINESTRING", "POINT"), function(x) st_cast(xg, x))
}
st_sfc(cast_all(mpl))
## no closing coordinates should remain for multipoint
any(duplicated(unclass(st_cast(mpl, "MULTIPOINT"))))  ## should be FALSE
## number of duplicated coordinates in the linestrings should equal the number of polygon rings 
## (... in this case, won't always be true)
sum(duplicated(do.call(rbind, unclass(st_cast(mpl, "MULTILINESTRING"))))) == sum(unlist(lapply(mpl, length)))  ## should be TRUE
mls <- st_cast(nc$geometry[[4]], "MULTILINESTRING")
st_sfc(cast_all(mls))
mpt <- st_cast(nc$geometry[[4]], "MULTIPOINT")
st_sfc(cast_all(mpt))
pl <- st_cast(nc$geometry[[4]], "POLYGON")
st_sfc(cast_all(pl))
ls <- st_cast(nc$geometry[[4]], "LINESTRING")
st_sfc(cast_all(ls))
pt <- st_cast(nc$geometry[[4]], "POINT")
## st_sfc(cast_all(pt))  ## Error: cannot create MULTIPOLYGON from POINT 
st_sfc(lapply(c("POINT", "MULTIPOINT"), function(x) st_cast(pt, x)))
}

