% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_sfc.R
\name{st_cast_sfc_default}
\alias{st_cast_sfc_default}
\title{Coerce geometry to MULTI* geometry}
\usage{
st_cast_sfc_default(x)
}
\arguments{
\item{x}{list of geometries or simple features}
}
\description{
Mixes of POINTS and MULTIPOINTS, LINESTRING and MULTILINESTRING, 
POLYGON and MULTIPOLYGON are returned as MULTIPOINTS, MULTILINESTRING and MULTIPOLYGONS respectively
}
\details{
Geometries that are already MULTI* are left unchanged. 
Features that can't be cast to a single  MULTI* geometry are return as a 
GEOMETRYCOLLECTION
}
