% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-predicates.R
\name{st_contains}
\alias{st_contains}
\alias{st_contains_properly}
\title{Identify if y is within x}
\usage{
st_contains(x, y, sparse = TRUE, prepared = TRUE, ..., model = "open")

st_contains_properly(x, y, sparse = TRUE, prepared = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{y}{object of class \code{sf}, \code{sfc} or \code{sfg}; if missing, \code{x} is used}

\item{sparse}{logical; should a sparse index list be returned (\code{TRUE}) or a dense logical matrix? See below.}

\item{prepared}{logical; prepare geometry for \code{x}, before looping over \code{y}? See Details.}

\item{model}{character; polygon/polyline model; one of
"open", "semi-open" or "closed"; see Details.}
}
\description{
\itemize{
\item \code{st_contains()} is true if
\item \code{st_contains_properly(x, y)} is true if \code{x} intersects \code{y}'s interior, but not its edges or exterior; \code{x} contains \code{x}, but \code{x} does not properly contain \code{x}.
}
}
\details{
for \code{model}, see https://github.com/r-spatial/s2/issues/32

See also \link{st_relate} and \url{https://en.wikipedia.org/wiki/DE-9IM} for a more detailed description of the underlying algorithms.
}
\seealso{
Other geometric binary predicates for two spatial objects: 
\code{\link{geos_binary_pred}},
\code{\link{st_equals}()},
\code{\link{st_intersects}()}
}
\concept{geometric binary predicates for two spatial objects}
