% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{st_crop}
\alias{st_crop}
\title{crop an sf object to a specific rectangle}
\usage{
st_crop(x, y, ...)
}
\arguments{
\item{x}{object of class \code{sf} or \code{sfc}}

\item{y}{numeric vector with named elements \code{xmin}, \code{ymin}, \code{xmax} and \code{ymax}, or object of class \code{bbox}}

\item{...}{ignored}
}
\description{
crop an sf object to a specific rectangle
}
