% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{st_drop_zm}
\alias{st_drop_zm}
\title{Drop Z and/or M dimensions from feature geometries}
\usage{
st_drop_zm(x, ...)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfc} or \code{sf}}

\item{...}{ignored}
}
\description{
Drop Z and/or M dimensions from feature geometries, resetting classes appropriately
}
\examples{
st_drop_zm(st_linestring(matrix(1:32,8)))
x = st_sfc(st_linestring(matrix(1:32,8)), st_linestring(matrix(1:8,2)))
st_drop_zm(x)
a = st_sf(a = 1:2, geom=x)
st_drop_zm(a)
}

