% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{st_erase_overlaps}
\alias{st_erase_overlaps}
\title{Erase overlaps}
\usage{
st_erase_overlaps(x)
}
\arguments{
\item{x}{object of class \code{sf} or \code{sfc}}
}
\value{
Object with overlaps erased.
}
\description{
Erase overlaps
}
\details{
Overlapping areas are erased from geometries that are indexed at greater
numbers in the argument to \code{x}. Furthermore, note that geometries that are empty 
or contained fully inside geometries with higher priority are removed entirely. 
See the \code{"idx"} attribute to find the orginal index for returned geometries.
}
\examples{
# example 1
## simulate data
pl1 = st_polygon(list(matrix(c(0, 0, 2, 0, 1, 1, 0 ,0), byrow = TRUE, ncol=2)))
pl2 = st_polygon(list(matrix(c(0, 0.5, 2, 0.5, 1, 1.5, 0, 0.5), byrow = TRUE, ncol = 2)))
pl3 = st_polygon(list(matrix(c(0, 1.25, 2, 1.25, 1, 2.5, 0, 1.25), byrow = TRUE, ncol = 2)))
x = st_sfc(list(pl1, pl2, pl3))
## erase overlaps
y = st_erase_overlaps(x)
## visualize differences
par(mfrow = c(1,2))
plot(x, main = "orginal data")
plot(y, main = "overlaps erased")
# example 2
## simulate data
pl1 = st_polygon(list(matrix(c(0, 0, 2, 0, 2, 2, 0, 2, 0, 0), byrow = TRUE, ncol=2)))
pl2 = st_polygon(list(matrix(c(0.5, 0.5, 1.5, 0.5, 1.5, 1.5, 0.5, 1.5, 0.5, 0.5), byrow = TRUE, ncol = 2)))
pl3 = st_polygon(list(matrix(c(5, 5, 7, 5, 7, 7, 5, 7, 5, 5), byrow = TRUE, ncol = 2)))
x = st_sfc(list(pl1, pl2, pl3))
## erase overlaps
y = st_erase_overlaps(x)
## visualize differences
par(mfrow = c(1,2))
plot(x, main = "orginal data")
plot(y, main = "overlaps erased")
## print the corresponding indices of y as they appeared in x
print(attr(y, "idx"))
}
