% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{st_geometry_type}
\alias{st_geometry_type}
\title{return geometry type of an object}
\usage{
st_geometry_type(x)
}
\arguments{
\item{x}{object of class \link{sf} or \link{sfc}}
}
\value{
returns a factor with the geometry type of each simple feature in x
}
\description{
return geometry type of an object, as a factor
}

