% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{st_is_full}
\alias{st_is_full}
\alias{st_is_full.sfg}
\alias{st_is_full.sfc}
\alias{st_is_full.sf}
\alias{st_is_full.bbox}
\title{predicate whether a geometry is equal to a POLYGON FULL}
\usage{
st_is_full(x, ...)

\method{st_is_full}{sfg}(x, ..., is_longlat = NULL)

\method{st_is_full}{sfc}(x, ...)

\method{st_is_full}{sf}(x, ...)

\method{st_is_full}{bbox}(x, ...)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfc} or \code{sf}}

\item{...}{ignored, except when it contains a \code{crs} argument to inform unspecified \code{is_longlat}}

\item{is_longlat}{logical; output of \link{st_is_longlat} of the parent \code{sfc} object}
}
\value{
logical, indicating whether geometries are POLYGON FULL (a spherical
polygon covering the entire sphere)
}
\description{
predicate whether a geometry is equal to a POLYGON FULL
}
