% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{st_is_longlat}
\alias{st_is_longlat}
\title{Assert whether simple feature coordinates are longlat degrees}
\usage{
st_is_longlat(x)
}
\arguments{
\item{x}{object of class \link{sf} or \link{sfc}}
}
\value{
TRUE if x has geographic coordinates
}
\description{
Assert whether simple feature coordinates are longlat degrees
}
