% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{st_list}
\alias{st_list}
\title{list layers in a datasource}
\usage{
st_list(dsn, options = character(0))
}
\arguments{
\item{dsn}{data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder, or contain the name and access credentials of a database)}

\item{options}{character; driver dependent dataset open options, multiple options supported.}
}
\description{
list layers in a datasource
}

