% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwgeom.R
\name{st_make_valid}
\alias{st_make_valid}
\title{try to make an invalid geometry valid}
\usage{
st_make_valid(x)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfg} or \code{sf}}
}
\value{
object of the same class as \code{x}
}
\description{
try to make an invalid geometry valid
}
\details{
uses the \code{lwgeom_makevalid} method also used by PostGIS' \code{ST_makevalid}.
}
