% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{st_makegrid}
\alias{st_makegrid}
\title{make a rectangular grid of polygons over the bounding box of a sf or sfc object}
\usage{
st_makegrid(x, cellsize = c(diff(st_bbox(x)[c(1, 3)]), diff(st_bbox(x)[c(2,
  4)]))/n, offset = st_bbox(x)[1:2], n = c(10, 10))
}
\arguments{
\item{x}{object of class \link{sf} or \link{sfc}}

\item{cellsize}{target cellsize}

\item{offset}{numeric of lengt 2; lower left corner coordinates (x, y) of the grid}

\item{n}{integer of length 1 or 2, number of grid cells in x and y direction (columns, rows)}
}
\description{
make a rectangular grid of polygons over the bounding box of a sf or sfc object
}

