% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{st_nearest_feature}
\alias{st_nearest_feature}
\title{get index of nearest feature}
\usage{
st_nearest_feature(x, y)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfc} or \code{sf}}

\item{y}{object of class \code{sfg}, \code{sfc} or \code{sf}}
}
\value{
for each feature (geometry) in \code{x} the index of the nearest feature (geometry) in \code{y}
}
\description{
get index of nearest feature
}
\examples{
ls1 = st_linestring(rbind(c(0,0), c(1,0)))
ls2 = st_linestring(rbind(c(0,0.1), c(1,0.1)))
ls3 = st_linestring(rbind(c(0,1), c(1,1)))
(l = st_sfc(ls1, ls2, ls3))

p1 = st_point(c(0.1, -0.1))
p2 = st_point(c(0.1, 0.11))
p3 = st_point(c(0.1, 0.09))
p4 = st_point(c(0.1, 0.9))

(p = st_sfc(p1, p2, p3, p4))
st_nearest_feature(p, l)
st_nearest_points(p, l[st_nearest_feature(p,l)], pairwise = TRUE)
}
