% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-measures.R, R/geom-transformers.R
\name{st_project_point}
\alias{st_project_point}
\alias{st_interpolate_line}
\title{Project point on linestring, interpolate along a linestring}
\usage{
st_project_point(line, point, normalized = FALSE)

st_interpolate_line(line, dist, normalized = FALSE)
}
\arguments{
\item{line}{object of class \code{sfc} with \code{LINESTRING} geometry}

\item{point}{object of class \code{sfc} with \code{POINT} geometry}

\item{normalized}{logical; if \code{TRUE}, use or return distance normalised to 0-1}

\item{dist}{numeric, vector with distance value(s)}
}
\description{
Project point on linestring, interpolate along a linestring
}
\examples{
st_project_point(st_as_sfc("LINESTRING (0 0, 10 10)"), st_as_sfc("POINT (5 5)"))
st_project_point(st_as_sfc("LINESTRING (0 0, 10 10)"), st_as_sfc("POINT (5 5)"), TRUE)
st_interpolate_line(st_as_sfc("LINESTRING (0 0, 1 1)"), 1)
st_interpolate_line(st_as_sfc("LINESTRING (0 0, 1 1)"), 1, TRUE)
}
