% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{st_read}
\alias{st_read}
\alias{st_read_pg}
\alias{st_write}
\title{read simple features from file or database}
\usage{
st_read(dsn, layer, ...)

st_write(sf, dsn = ".", layer, driver = "ESRI Shapefile",
  opts = character(), ...)

st_read_pg(cn = NULL, query, dbname, geom_column = NULL, ...)
}
\arguments{
\item{dsn}{data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder)}

\item{layer}{layer name (varies by driver, may be a file name without extension)}

\item{...}{parameter(s) passed on to \link{st_as_sf}}

\item{sf}{object of class \code{sf}}

\item{driver}{driver name}

\item{opts}{options to pass on to driver}

\item{cn}{open database connection}

\item{query}{SQL query to select records}

\item{dbname}{character; database name, only used if cn is \code{NULL}}

\item{geom_column}{character or integer: indicator of name or position of the geometry column; if not provided, the last column of type character is chosen}
}
\description{
read simple features from file or database

read PostGIS table directly through DBI and RPostgreSQL interface, converting wkb
}
\examples{
if (Sys.getenv("USER") == "edzer") {
 (s = st_read("PG:dbname=test", "test"))
 summary(s)
}
(s = st_read(system.file("shapes/", package="maptools"), "sids"))[1:10,]
summary(s)
if (Sys.getenv("USER") == "edzer") {
  st_read_pg(dbname = "postgis", query = "select * from meuse2 limit 3;")
}
}

