% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recenter.R
\name{st_recenter}
\alias{st_recenter}
\alias{st_recenter.sfc}
\alias{st_recenter.sf}
\title{Shift or re-center geographical coordinates for a Pacific view}
\usage{
st_recenter(x)

\method{st_recenter}{sfc}(x, ...)

\method{st_recenter}{sf}(x, ...)
}
\arguments{
\item{x}{object of class sf or sfc}

\item{...}{ignored}
}
\description{
All longitudes < 0 are added to 360, to avoid for instance parts of Alaska
  being represented on the far left and right of a plot because they have
  values straddling 180 degrees. In general, using a projected
  coordinate reference system is to be preferred, but this method permits a
  geographical coordinate reference system to be used. This is the sf
  equivalent of \code{\link[sp]{recenter}}.
}
\examples{
## sfc
pt1 = st_point(c(-170, 50))
pt2 = st_point(c(170, 50))
(sfc = st_sfc(pt1, pt2))
sfc = st_set_crs(sfc, 4326)
st_recenter(sfc)

## sf
d = st_as_sf(data.frame(id = 1:2, geometry = sfc))
st_recenter(d)
}
