% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{ST_sfc}
\alias{ST_sfc}
\title{verify simple feature}
\usage{
ST_sfc(lst, epsg = NA_integer_, proj4string = NA_character_)
}
\arguments{
\item{lst}{list with simple feature objects}

\item{epsg}{integer; epsg code}

\item{proj4string}{character; describing the coordinate reference systems in PROJ.4 syntax}
}
\description{
verifies simple feature list column's contents, and sets class
}
\examples{
pt1 = ST_Point(c(0,1))
pt2 = ST_Point(c(1,1))
(sfc = ST_sfc(list(pt1, pt2)))
d = data.frame(a = 1:2)
}

