% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{st_transform}
\alias{st_transform}
\alias{st_transform.sf}
\alias{st_transform.sfc}
\alias{st_transform.sfi}
\title{coordinate transformation or conversion}
\usage{
st_transform(x, proj4string)

\method{st_transform}{sfc}(x, proj4string = NA_character_, ...)

\method{st_transform}{sf}(x, proj4string = NA_character_, ...)

\method{st_transform}{sfi}(x, proj4string = NA_character_, ...)
}
\arguments{
\item{x}{object of class sf, sfc or sfi}

\item{proj4string}{character; describing the coordinate reference systems in PROJ.4 syntax}

\item{...}{ignored}
}
\description{
coordinate transformation or conversion
}
\details{
transforms coordinates of object to new projection

the st_transform method for sfi objects assumes that the proj4string of the object is available as an attribute of that name.
}
\examples{
p1 = st_point(c(7,52))
p2 = st_point(c(-30,20))
sfc = st_sfc(p1, p2, proj4string = "+init=epsg:4326")
st_transform(sfc, "+init=epsg:3857")
st_transform(st_sf(a=2:1, geom=sfc), "+init=epsg:3857")
st_transform(structure(p1, proj4string = "+init=epsg:4326"), "+init=epsg:3857")
}

