% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{st_write_db}
\alias{st_write_db}
\title{write simple feature table to a spatial database}
\usage{
st_write_db(conn = NULL, obj, table_name = substitute(obj),
  geom_name = "wkb_geometry", ..., dropTable = FALSE, wkb = TRUE)
}
\arguments{
\item{conn}{open database connection}

\item{obj}{object of class \code{sf}}

\item{table_name}{name for the table in the database}

\item{geom_name}{name of the geometry column in the database}

\item{...}{arguments passed on to \code{dbWriteTable}}

\item{dropTable}{logical; should \code{table_name} be dropped first?}

\item{wkb}{logical; use well-known-binary for transfer?}
}
\description{
write simple feature table to a spatial database
}
\examples{
library(sp)
data(meuse)
sf = st_as_sf(meuse, coords = c("x", "y"), epsg = 28992)
library(RPostgreSQL)
conn = dbConnect(PostgreSQL(), dbname = "postgis")
st_write_db(conn, sf, "meuse", dropTable = FALSE)
}

