% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbox.R
\docType{data}
\name{st_zbox}
\alias{st_zbox}
\alias{is.na.zbox}
\alias{st_zbox.POINT}
\alias{st_zbox.MULTIPOINT}
\alias{st_zbox.LINESTRING}
\alias{st_zbox.POLYGON}
\alias{st_zbox.MULTILINESTRING}
\alias{st_zbox.MULTIPOLYGON}
\alias{st_zbox.GEOMETRYCOLLECTION}
\alias{st_zbox.MULTISURFACE}
\alias{st_zbox.MULTICURVE}
\alias{st_zbox.CURVEPOLYGON}
\alias{st_zbox.COMPOUNDCURVE}
\alias{st_zbox.POLYHEDRALSURFACE}
\alias{st_zbox.TIN}
\alias{st_zbox.TRIANGLE}
\alias{st_zbox.CIRCULARSTRING}
\alias{st_zbox.sfc}
\alias{st_zbox.sf}
\alias{st_zbox.numeric}
\alias{NA_zbox_}
\title{Return bounding of a simple feature or simple feature set}
\format{An object of class \code{zbox} of length 2.}
\usage{
\method{is.na}{zbox}(x)

st_zbox(obj, ...)

\method{st_zbox}{POINT}(obj, ...)

\method{st_zbox}{MULTIPOINT}(obj, ...)

\method{st_zbox}{LINESTRING}(obj, ...)

\method{st_zbox}{POLYGON}(obj, ...)

\method{st_zbox}{MULTILINESTRING}(obj, ...)

\method{st_zbox}{MULTIPOLYGON}(obj, ...)

\method{st_zbox}{GEOMETRYCOLLECTION}(obj, ...)

\method{st_zbox}{MULTISURFACE}(obj, ...)

\method{st_zbox}{MULTICURVE}(obj, ...)

\method{st_zbox}{CURVEPOLYGON}(obj, ...)

\method{st_zbox}{COMPOUNDCURVE}(obj, ...)

\method{st_zbox}{POLYHEDRALSURFACE}(obj, ...)

\method{st_zbox}{TIN}(obj, ...)

\method{st_zbox}{TRIANGLE}(obj, ...)

\method{st_zbox}{CIRCULARSTRING}(obj, ...)

\method{st_zbox}{sfc}(obj, ...)

\method{st_zbox}{sf}(obj, ...)

\method{st_zbox}{numeric}(obj, ..., crs = NA_crs_)

NA_zbox_
}
\arguments{
\item{x}{object of class \code{zbox}}

\item{obj}{object to compute the bounding box from}

\item{...}{ignored}

\item{crs}{object of class \code{crs}, or argument to \link{st_crs}, specifying the CRS of this bounding box.}
}
\value{
a numeric vector of length two, with \code{zmin} and \code{zmax} values;
if \code{obj} is of class \code{sf} or \code{sfc} the object
returned has a class \code{zbox}, an attribute \code{crs} and a method to print the
zbox and an \code{st_crs} method to retrieve the coordinate reference system
corresponding to \code{obj} (and hence the bounding box). \link{st_as_sfc} has a
methods for \code{zbox} objects to generate a polygon around the four bounding box points.
}
\description{
Return bounding of a simple feature or simple feature set
}
\details{
\code{NA_zbox_} represents the missing value for a \code{zbox} object
}
\examples{
a = st_sf(a = 1:2, geom = st_sfc(st_point(0:1), st_point(1:2)), crs = 4326)
st_zbox(a)
st_as_sfc(st_zbox(a))
st_zbox(c(zmin = 16.1, zmax = 16.6), crs = st_crs(4326))
}
\keyword{datasets}
