% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{summary.sfc}
\alias{summary.sfc}
\title{summarize simple feature column}
\usage{
\method{summary}{sfc}(object, ..., maxsum = 7L, maxp4s = 10L)
}
\arguments{
\item{object}{object of class \code{sfc}}

\item{...}{ignored}

\item{maxsum}{maximum number of classes to summarize the simple feature column to}

\item{maxp4s}{maximum number of characters to print from the PROJ.4 string}
}
\description{
summarize simple feature column
}

