% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{vctrs}
\alias{vctrs}
\alias{vec_ptype2.sfc}
\alias{vec_ptype2.sfc.default}
\alias{vec_ptype2.sfc.sfc}
\alias{vec_cast.sfc}
\alias{vec_cast.sfc.sfc}
\alias{vec_cast.sfc.default}
\title{vctrs methods for sf objects}
\usage{
vec_ptype2.sfc(x, y, ...)

\method{vec_ptype2.sfc}{default}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2.sfc}{sfc}(x, y, ...)

vec_cast.sfc(x, to, ...)

\method{vec_cast.sfc}{sfc}(x, to, ...)

\method{vec_cast.sfc}{default}(x, to, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}

\item{x_arg}{Argument names for \code{x} and \code{y}. These are used
in error messages to inform the user about the locations of
incompatible types (see \code{\link[=stop_incompatible_type]{stop_incompatible_type()}}).}

\item{y_arg}{Argument names for \code{x} and \code{y}. These are used
in error messages to inform the user about the locations of
incompatible types (see \code{\link[=stop_incompatible_type]{stop_incompatible_type()}}).}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}
}
\description{
vctrs methods for sf objects
}
