% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.sf}
\alias{aggregate.sf}
\title{aggregate an \code{sf} object}
\usage{
\method{aggregate}{sf}(x, by, FUN, ..., do_union = TRUE, simplify = TRUE)
}
\arguments{
\item{x}{object of class \link{sf}}

\item{by}{see \link[stats]{aggregate}}

\item{FUN}{function passed on to \link[stats]{aggregate}, in case \code{ids} was specified and attributes need to be grouped}

\item{...}{arguments passed on to \code{FUN}}

\item{do_union}{logical; should grouped geometries be unioned using \link{st_union}?}

\item{simplify}{logical; see \link[stats]{aggregate}}
}
\value{
an \code{sf} object with aggregated attributes and geometries, with an additional grouping variable called \code{Group.1}.
}
\description{
aggregate an \code{sf} object, possibly union-ing geometries
}
