% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{methods}
\name{dbDataType,PostgreSQLConnection,sf-method}
\alias{dbDataType,PostgreSQLConnection,sf-method}
\title{Determine database type for R vector}
\usage{
\S4method{dbDataType}{PostgreSQLConnection,sf}(dbObj, obj)
}
\arguments{
\item{dbObj}{PostgreSQLConnection driver or connection.}

\item{obj}{Object to convert}
}
\description{
Determine database type for R vector
}
