% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{methods}
\name{dbDataType,PostgreSQLConnection,sf-method}
\alias{dbDataType,PostgreSQLConnection,sf-method}
\alias{dbDataType,DBIObject,sf-method}
\title{Determine database type for R vector}
\usage{
\S4method{dbDataType}{PostgreSQLConnection,sf}(dbObj, obj)

\S4method{dbDataType}{DBIObject,sf}(dbObj, obj)
}
\arguments{
\item{dbObj}{DBIObject driver or connection.}

\item{obj}{Object to convert}
}
\description{
Determine database type for R vector

Determine database type for R vector
}
