% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{interpolate_aw}
\alias{interpolate_aw}
\alias{st_interpolate_aw}
\alias{st_interpolate_aw.sf}
\alias{st_interpolate_aw.stars}
\title{Area-weighted interpolation or dasymetric mapping of polygon data}
\usage{
st_interpolate_aw(x, to, extensive, ...)

\method{st_interpolate_aw}{sf}(
  x,
  to,
  extensive,
  ...,
  keep_NA = FALSE,
  na.rm = FALSE,
  include_non_intersected = FALSE,
  weights = character(0)
)

\method{st_interpolate_aw}{stars}(x, to, extensive, ...)
}
\arguments{
\item{x}{object of class \code{sf}, for which we want to aggregate attributes}

\item{to}{object of class \code{sf} or \code{sfc}, with the target geometries}

\item{extensive}{logical; if TRUE, the attribute variables are assumed to be spatially extensive (like population) and the sum is preserved, otherwise, spatially intensive (like population density) and the mean is preserved.}

\item{...}{ignored}

\item{keep_NA}{logical; if \code{TRUE}, return all features in \code{to}, if \code{FALSE} return only those with non-NA values (but with \code{row.names} the index corresponding to the feature in \code{to})}

\item{na.rm}{logical; if \code{TRUE} remove features with \code{NA} attributes from \code{x} before interpolating}

\item{include_non_intersected}{logical; for the case when \code{extensive=FALSE}, when set to \code{TRUE} divide by the target areas (including non-intersected areas),
when \code{FALSE} divide by the sum of the source areas.}

\item{weights}{character; name of column in \code{to} that indicates (extensive) weights, to be used instead of areas, for redistributing attributes in \code{x}; currently only works for \code{extensive=TRUE}.}
}
\description{
Area-weighted interpolation or dasymetric mapping of polygon data
}
\details{
if \code{extensive} is \code{TRUE} and \code{na.rm} is set to \code{TRUE}, geometries with \code{NA} are effectively treated as having zero attribute values. Dasymetric mapping is obtained when \code{weights} are specified.
}
\examples{
# example Area-weighted interpolation:
nc = st_read(system.file("shape/nc.shp", package="sf"))
g = st_make_grid(nc, n = c(10, 5))
a1 = st_interpolate_aw(nc["BIR74"], g, extensive = FALSE)
sum(a1$BIR74) / sum(nc$BIR74) # not close to one: property is assumed spatially intensive
a2 = st_interpolate_aw(nc["BIR74"], g, extensive = TRUE)
# verify mass preservation (pycnophylactic) property:
sum(a2$BIR74) / sum(nc$BIR74)
a1$intensive = a1$BIR74
a1$extensive = a2$BIR74
\donttest{plot(a1[c("intensive", "extensive")], key.pos = 4)}

# example Dasymetric mapping:
# load nr of addresses per 10 km grid cell, to proxy population -> birth density:
grd.addr = system.file("gpkg/grd_addr.gpkg", package="sf") |> read_sf()
nc = st_transform(nc, st_crs(grd.addr))
# avoid "assumes attributes are constant or uniform over areas" warnings:
st_agr(grd.addr) = c(ones = "constant")
st_agr(nc) = c(BIR74 = "constant", BIR79 = "constant")
\donttest{plot(nc["BIR74"], logz = TRUE, main = "county birth counts, 1974-")}
bir0.grd = st_interpolate_aw(nc[c("BIR74","BIR79")], extensive = TRUE, grd.addr)
\donttest{plot(bir0.grd["BIR74"], logz = TRUE, main = "area-weighted birth counts, 1974-")}
xgrd.addr = grd.addr # copy for plotting
xgrd.addr$ones[grd.addr$ones==0] = 1 # so that logz shows finite values
\donttest{plot(xgrd.addr, logz = TRUE, main = "nr of addresses per cell") # log scale}
# dasymetric mapping
bir.grd = st_interpolate_aw(nc[c("BIR74","BIR79")], extensive = TRUE, grd.addr, weights = "ones")
xbir.grd = bir.grd # copy for plotting
smallest_pos = function(x) min(x[x > 0])
xbir.grd$BIR74[xbir.grd$BIR74 == 0] = smallest_pos(xbir.grd$BIR74) # again bc of logz
\donttest{plot(xbir.grd["BIR74"], logz = TRUE, main = "dasym. redistributed birth counts, 1974-")}
# verify sums:
apply(as.data.frame(bir.grd)[1:2], 2, sum)
apply(as.data.frame(nc)[c("BIR74", "BIR79")], 2, sum)
# compare county-wise:
st_agr(bir.grd) = c(BIR74 = "constant")
\donttest{aw <- st_interpolate_aw(bir.grd["BIR74"], st_geometry(nc), extensive = TRUE)}
\donttest{plot(nc$BIR74, aw$BIR74, log = 'xy', xlab = 'county-value', ylab = 'area-w interpolated')}
\donttest{abline(0,1)}
}
