% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nc}
\alias{nc}
\title{North Carolina counties SIDS dataset}
\description{
North Carolina counties SIDS dataset
}
\seealso{
\url{https://cran.r-project.org/web/packages/spdep/vignettes/sids.pdf}
}
\keyword{data}
