% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nc}
\alias{nc}
\title{North Carolina SIDS data}
\format{
A \code{sf} object
}
\description{
Sudden Infant Death Syndrome (SIDS) sample data for North Carolina counties,
two time periods (1974-78 and 1979-84). The details of the columns can be
found in a \href{https://r-spatial.github.io/spdep/articles/sids.html}{spdep package vignette}.
Please note that, though this is basically the same as \code{nc.sids} dataset in spData
package, \code{nc} only contains a subset of variables. The differences are
also discussed on the vignette.
}
\examples{
\donttest{
nc <- st_read(system.file("shape/nc.shp", package="sf"))
}
}
\seealso{
\url{https://r-spatial.github.io/spdep/articles/sids.html}
}
