% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{s2}
\alias{s2}
\alias{st_as_sfc.s2latlng}
\alias{st_as_sfc.s2polyline}
\alias{st_as_sfc.s2polygon}
\alias{s2latlng.sfc}
\alias{s2polyline.sfc}
\alias{s2polygon.sfc}
\alias{st_as_s2}
\alias{st_as_s2.sfc_MULTIPOLYGON}
\alias{st_as_s2.sfc_POLYGON}
\alias{st_as_s2.sfc_POINT}
\alias{st_as_s2.sfc_LINESTRING}
\alias{st_as_s2.sfc_MULTILINESTRING}
\alias{st_as_s2.sfc}
\alias{st_as_s2.sf}
\alias{st_as_s2.sfg}
\alias{st_as_sfc.S2Polyline}
\alias{st_as_sfc.S2Point}
\alias{st_intersects.S2Polygon}
\alias{st_is_valid.S2Polygon}
\alias{st_is_valid.S2Polyline}
\alias{st_area.S2Polygon}
\title{functions for spherical geometry, using libs2 package}
\usage{
\method{st_as_sfc}{s2latlng}(x, ...)

\method{st_as_sfc}{s2polyline}(x, ...)

\method{st_as_sfc}{s2polygon}(x, ...)

s2latlng.sfc(x, ...)

s2polyline.sfc(x, ...)

s2polygon.sfc(x, ..., oriented = FALSE, check = TRUE)

st_as_s2(x, ...)

\method{st_as_s2}{sfc_MULTIPOLYGON}(x, ..., oriented = FALSE, check = TRUE)

\method{st_as_s2}{sfc_POLYGON}(x, ..., oriented = FALSE, check = TRUE)

\method{st_as_s2}{sfc_POINT}(x, ..., oriented = FALSE)

\method{st_as_s2}{sfc_LINESTRING}(x, ..., oriented = FALSE)

\method{st_as_s2}{sfc_MULTILINESTRING}(x, ...)

\method{st_as_s2}{sfc}(x, ...)

\method{st_as_s2}{sf}(x, ..., oriented = FALSE, check = TRUE)

\method{st_as_s2}{sfg}(x, ..., oriented = FALSE, check = TRUE, crs = st_crs("EPSG:4326"))

\method{st_as_sfc}{S2Polyline}(x, ...)

\method{st_as_sfc}{S2Point}(x, ...)

\method{st_intersects}{S2Polygon}(x, y = x, sparse = TRUE, ...)

\method{st_is_valid}{S2Polygon}(x, ...)

\method{st_is_valid}{S2Polyline}(x, ...)

\method{st_area}{S2Polygon}(x, ..., radius = units::set_units(6371008.8, "m", mode = "standard"))
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{...}{passed on; might contain named argument crs}

\item{oriented}{logical, see \link[libs2]{s2polygon}}

\item{check}{logical, see \link[libs2]{s2polygon}}

\item{crs}{object of class \code{crs}}

\item{y}{object of class \code{S2Polygon}}

\item{sparse}{logical; see \link{st_intersects}}

\item{radius}{numeric or \code{units}; radius of the (spherical) Earth.}
}
\description{
functions for spherical geometry, using the libs2 package based on the google s2geometry.io library
}
\details{
\code{st_as_s2} converts an \code{sf} POLYGON object into a form readable by \code{s2}.
}
\examples{
m = rbind(c(-1,-1), c(1,-1), c(1,1), c(-1,1), c(-1,-1))
m1 = rbind(c(-1,-1), c(1,-1), c(1,1), c(-1,1), c(-1,0), c(-1,-1))
m0 = m[5:1,]
mp = st_multipolygon(list(
list(m, 0.8 * m0, 0.01 * m1 + 0.9),
list(0.7* m, 0.6*m0), 
list(0.5 * m0), 
list(m+2),
list(m+4,(.9*m0)+4)
))
sf = st_sfc(mp, mp, crs = 'EPSG:4326')
s2 = st_as_s2(sf)
s2pts = st_as_s2(st_sfc(st_point(c(0,0)), st_point(c(2,2))), crs = 4326)
ls = st_sfc(st_linestring(rbind(c(2,1), c(1,2), c(0,0), c(1,1))), crs = 4326)
s2ls = st_as_s2(ls)
mls = st_sfc(st_multilinestring(list(rbind(c(2,1), c(1,2)), rbind(c(0,0), c(1,1)))), crs = 4326)
try(st_as_s2(mls))
try(st_as_s2(st_multipoint(rbind(c(0,0), c(1,1)))))
st_as_sfc(s2ls)
st_as_sfc(s2pts)
nc = st_read(system.file("shape/nc.shp", package="sf"))
st_is_valid(st_as_s2(nc[1:10,]))
st_is_valid(s2ls)
st_area(st_as_s2(nc[1:10,]))
st_area(nc[1:10,])/st_area(st_as_s2(nc[1:10,]))
}
