% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{s2}
\alias{s2}
\alias{st_as_s2}
\alias{st_as_s2.sfc}
\alias{st_as_sfc.S2Polygon}
\alias{st_intersects.S2Polygon}
\alias{st_is_valid.S2Polygon}
\alias{st_area.S2Polygon}
\title{functions for spherical geometry, using libs2 package}
\usage{
st_as_s2(x, ...)

\method{st_as_s2}{sfc}(x, ..., oriented = FALSE)

\method{st_as_sfc}{S2Polygon}(x, ...)

\method{st_intersects}{S2Polygon}(x, y = x, sparse = TRUE, ...)

\method{st_is_valid}{S2Polygon}(x, ...)

\method{st_area}{S2Polygon}(x, ..., radius = units::set_units(6371008.8, "m", mode = "standard"))
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{...}{ignored}

\item{oriented}{logical; are the polygon rings guaranteed to be oriented (CCW outer rings, CW holes)?}

\item{y}{object of class \code{S2Polygon}}

\item{sparse}{logical; see \link{st_intersects}}

\item{radius}{numeric or \code{units}; radius of the (spherical) Earth.}
}
\description{
functions for spherical geometry, using the libs2 package based on the google s2geometry.io library
}
\details{
\code{st_as_s2} converts an \code{sf} POLYGON object into a form readable by \code{s2}.
}
\examples{
m = rbind(c(-1,-1), c(1,-1), c(1,1), c(-1,1), c(-1,-1))
m1 = rbind(c(-1,-1), c(1,-1), c(1,1), c(-1,1), c(-1,0), c(-1,-1))
m0 = m[5:1,]
mp = st_multipolygon(list(
list(m, 0.8 * m0, 0.01 * m1 + 0.9),
list(0.7* m, 0.6*m0), 
list(0.5 * m0), 
list(m+2),
list(m+4,(.9*m0)+4)
))
sf = st_sfc(mp, mp, crs = 'EPSG:4326')
s2 = st_as_s2(sf)
x = st_as_sfc(s2)
all.equal(st_area(sf), st_area(x))
nc = st_read(system.file("shape/nc.shp", package="sf"))
st_is_valid(st_as_s2(nc[1:10,]))
st_area(st_as_s2(nc[1:10,]))
st_area(nc[1:10,])/st_area(st_as_s2(nc[1:10,]))
}
