% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{s2}
\alias{s2}
\alias{st_as_s2}
\alias{s2}
\alias{st_as_sfc.S2Polygon}
\alias{s2}
\alias{s2_intersection}
\alias{s2}
\alias{s2_intersects}
\alias{s2}
\alias{s2_centroid}
\alias{s2}
\alias{s2_area}
\title{functions for spherical geometry, using s2 package}
\usage{
st_as_s2(x)

\method{st_as_sfc}{S2Polygon}(x, ..., crs = st_crs(4326))

s2_intersection(x, y)

s2_intersects(x, y)

s2_centroid(x)

s2_area(x)
}
\arguments{
\item{x}{object of class \code{sf}, or \code{sfc}}

\item{...}{passed on to \link{st_transform}}

\item{crs}{object of class \code{crs}}

\item{y}{object of class \code{sf}, or \code{sfc}}
}
\description{
functions for spherical geometry, using the s2 package based on the google s2 library
}
\details{
\code{st_as_s2} converts an \code{sf} POLYGON object into a form readable by \code{s2}.

\code{s2_centroid} computes the spherical centroid of a set of polygons

\code{s2_area} computes the area of a set of polygons, as a fraction of 4 * pi.
}
