% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\docType{package}
\name{sf-defunct}
\alias{sf-defunct}
\alias{st_read_db}
\alias{st_read_db,}
\alias{st_write_db}
\title{Deprecated functions in \code{sf}}
\usage{
st_read_db(
  conn = NULL,
  table = NULL,
  query = NULL,
  geom_column = NULL,
  EWKB = TRUE,
  ...
)
}
\arguments{
\item{conn}{open database connection}

\item{table}{table name}

\item{query}{SQL query to select records; see details}

\item{geom_column}{deprecated. Geometry column name}

\item{EWKB}{logical; is the WKB of type EWKB? if missing, defaults to
\code{TRUE}}

\item{...}{parameter(s) passed on to \link{st_as_sf}}
}
\description{
These functions are provided for compatibility with older version of \code{sf}.
They may eventually be completely removed.
}
\details{
The \code{geom_column} argument is deprecated. The function will
automatically find the \code{geometry} type columns. For the \code{RPostgreSQL} drivers
it will try to cast all the character columns, which can be long for very wide
tables.
}
\section{Details}{

\tabular{rl}{
\code{st_read_db} \tab now a synonym for \code{\link{st_read}}\cr
\code{st_write_db} \tab now a synonym for \code{\link{st_write}}\cr
}
}

