% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{sf-deprecated}
\alias{sf-deprecated}
\alias{st_read_db}
\alias{st_read_db,}
\alias{st_write_db}
\alias{sf-deprecated-package}
\title{Deprecated functions in \code{sf}}
\usage{
st_read_db(conn = NULL, table = NULL, query = NULL, geom_column = NULL,
  EWKB = TRUE, ...)
}
\arguments{
\item{conn}{open database connection}

\item{table}{table name}

\item{query}{SQL query to select records; see details}

\item{geom_column}{character or integer: indicator of name or position of the geometry column; if not provided, the last column of type character is chosen}

\item{EWKB}{logical; is the WKB is of type EWKB? if missing, defaults to \code{TRUE}}

\item{...}{parameter(s) passed on to \link{st_as_sf}}
}
\description{
These functions are provided for compatibility with older version of \code{sf}.
They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
\code{st_read_db} \tab now a synonym for \code{\link{st_read}}\cr
\code{st_write_db} \tab now a synonym for \code{\link{st_write}}\cr
}
}

