% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-package.R
\docType{package}
\name{sf-package}
\alias{sf-package}
\title{sf: Simple Features for R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Support for simple feature access, a standardized way to encode and analyze spatial vector data. Binds to 'GDAL' \doi{ 10.5281/zenodo.5884351} for reading and writing data, to 'GEOS' \doi{ 10.5281/zenodo.11396894} for geometrical operations, and to 'PROJ' \doi{ 10.5281/zenodo.5884394} for projection conversions and datum transformations. Uses by default the 's2' package for geometry operations on geodetic (long/lat degree) coordinates.
}
\references{
Pebesma, E. and Bivand, R. (2023). Spatial Data
Science: With Applications in R. Chapman and Hall/CRC.
\doi{10.1201/9780429459016} which is also found freely
online at \url{https://r-spatial.org/book/}

Pebesma, E., 2018. Simple Features for R: Standardized Support
for Spatial Vector Data. The R Journal 10 (1), 439-446,
\doi{10.32614/RJ-2018-009} (open access)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-spatial.github.io/sf/}
  \item \url{https://github.com/r-spatial/sf}
  \item Report bugs at \url{https://github.com/r-spatial/sf/issues}
}

}
\author{
\strong{Maintainer}: Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})

Other contributors:
\itemize{
  \item Roger Bivand (\href{https://orcid.org/0000-0003-2392-6140}{ORCID}) [contributor]
  \item Etienne Racine [contributor]
  \item Michael Sumner [contributor]
  \item Ian Cook [contributor]
  \item Tim Keitt [contributor]
  \item Robin Lovelace [contributor]
  \item Hadley Wickham [contributor]
  \item Jeroen Ooms (\href{https://orcid.org/0000-0002-4035-0289}{ORCID}) [contributor]
  \item Kirill Müller [contributor]
  \item Thomas Lin Pedersen [contributor]
  \item Dan Baston [contributor]
  \item Dewey Dunnington (\href{https://orcid.org/0000-0002-9415-4582}{ORCID}) [contributor]
}

}
\keyword{internal}
