% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf}
\alias{ST_sf}
\alias{[.sf}
\alias{p4s}
\alias{sf}
\title{create sf object}
\usage{
ST_sf(..., relation_to_geometry = NA_character_, row.names,
  stringsAsFactors = default.stringsAsFactors())

\method{[}{sf}(x, i, j, drop)

p4s(obj)
}
\arguments{
\item{...}{column elements to be binded into an \code{sf} object, one of them being of class \code{sfc}}

\item{relation_to_geometry}{character vector; see details below.}

\item{row.names}{row.names for the created \code{sf} object}

\item{stringsAsFactors}{logical; see \link{data.frame}}

\item{x}{object of class \code{sf}}

\item{i}{record selection, see \link{[.data.frame}}

\item{j}{variable selection, see \link{[.data.frame}}

\item{drop}{whether to drop to simpler (e.g. vector) representation, see \link{[.data.frame}}

\item{obj}{object of class \code{sf} or \code{sfc}}
}
\description{
create sf, which extends data.frame-like objects with a simple feature list column
}
\details{
\code{relation_to_geometry} specified for each non-geometry column how it relates to the geometry, and can have one of following values: "field", "lattice", "entity". "field" is used for attributes that are constant throughout the geometry (e.g. land use), "lattice" where the attribute is an aggregate value over the geometry (e.g. population density), "entity" when the attributes identifies the geometry of particular "thing", such as a building or a city. The default value, \code{NA_character_}, implies we don't know.

"[.sf" will return a \code{data.frame} if the geometry column (of class \code{sfc}) is dropped, an \code{sfc} object if only the geometry column is selected, otherwise the behavior depending on \code{drop} is identical to that of \link{[.data.frame}.

\code{p4s} returns the PROJ.4 string; if an EPSG code is available, it constructs it from this, otherwise, it takes the \code{proj4string} attribute, if none of these is available (both are missing-valued), \code{NULL} is returned.
}
\examples{
g = ST_sfc(list(ST_Point(1:2)))
ST_sf(a=3,g)
ST_sf(g, a=3)
ST_sf(a=3, ST_sfc(list(ST_Point(1:2)))) # better to name it!
}

