% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{sfc}
\alias{sfc}
\alias{st_sfc}
\title{create simple feature collection object of class sfc from list}
\usage{
st_sfc(..., crs = NA_integer_, precision = 0)
}
\arguments{
\item{...}{one or more simple feature objects}

\item{crs}{coordinate reference system: integer with the epsg code, or character with proj4string}

\item{precision}{numeric; see \link{st_as_binary}}
}
\description{
create simple feature list column, set class, and add coordinate reference system
}
\details{
a simple feature collection object is a list of class \code{c("stc_TYPE", "sfc")} which contains objects of identical type. This function creates such an object from a list of simple feature objects (of class \code{sfg}), and coerces their type if necessary: collections of XX and MULTIXX are coerced to MULTIXX (with XX: POINT, LINESTRING or POLYGON), other sets are coerced to GEOMETRYCOLLECTION.

in case \code{epsg} is given but \code{proj4string} is not and packages \code{sp} and \code{rgdal} can be loaded, the \code{proj4string} is expanded using the PROJ.4 epsg database.
}
\examples{
pt1 = st_point(c(0,1))
pt2 = st_point(c(1,1))
(sfc = st_sfc(pt1, pt2))
d = data.frame(a = 1:2)
}

