% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgbp.R
\name{sgbp}
\alias{sgbp}
\alias{print.sgbp}
\alias{sgbp}
\alias{t.sgbp}
\title{Methods for dealing with sparse geometry binary predicate lists}
\usage{
\method{print}{sgbp}(x, ..., n = 10, max_nb = 10)

\method{t}{sgbp}(x)
}
\arguments{
\item{x}{object of class \code{sgbp}}

\item{...}{ignored}

\item{n}{integer; maximum number of items to print}

\item{max_nb}{integer; maximum number of neighbours to print for each item}
}
\description{
Methods for dealing with sparse geometry binary predicate lists
}
\details{
\code{sgbp} are sparse matrices, stored as a list with integer vectors holding the \code{TRUE} indices of each row. This means that for a dense, \eqn{m \times n}{m x n} matrix \code{Q} and a list \code{L}, if \code{Q[i,j]} is \code{TRUE} then \eqn{j} is an element of \code{L[[i]]}. Reversed: when \eqn{k} is the value of \code{L[[i]][j]}, then \code{Q[i,k]} is \code{TRUE}.
}
