% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R, R/wkb.R
\name{ST_as.sfc}
\alias{ST_as.sfc}
\alias{ST_as.sfc.SpatialLines}
\alias{ST_as.sfc.SpatialMultiPoints}
\alias{ST_as.sfc.SpatialPixels}
\alias{ST_as.sfc.SpatialPoints}
\alias{ST_as.sfc.SpatialPolygons}
\alias{ST_as.sfc.WKB}
\title{convert foreign geometry object to an sfc object}
\usage{
ST_as.sfc(x, ...)

\method{ST_as.sfc}{SpatialPoints}(x, ...)

\method{ST_as.sfc}{SpatialPixels}(x, ...)

\method{ST_as.sfc}{SpatialMultiPoints}(x, ...)

\method{ST_as.sfc}{SpatialLines}(x, ..., forceMulti = FALSE)

\method{ST_as.sfc}{SpatialPolygons}(x, ..., forceMulti = FALSE)

\method{ST_as.sfc}{WKB}(x, ..., EWKB = FALSE)
}
\arguments{
\item{x}{object to convert}

\item{...}{further arguments}

\item{forceMulti}{logical; if \code{TRUE}, force coercion into \code{MULTIPOLYGON} or \code{MULTILINE} objects, else autodetect}

\item{EWKB}{logical; if TRUE, parse as EWKB (PostGIS: ST_AsEWKB), otherwise as ISO WKB (PostGIS: ST_AsBinary)}
}
\description{
convert foreign geometry object to an sfc object
}

