% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{ST_as.WKT}
\alias{ST_as.WKT}
\alias{ST_as.WKT.sfc}
\alias{ST_as.WKT.sfi}
\title{Return Well-known Text representation of simple feature item}
\usage{
ST_as.WKT(x, ...)

\method{ST_as.WKT}{sfi}(x, ...)

\method{ST_as.WKT}{sfc}(x, ...)
}
\arguments{
\item{x}{object of class sfi or sfc}

\item{...}{passed on to WKT_name}
}
\description{
Return Well-known Text representation of simple feature item
}
\details{
to suppress printing of SRID, \code{EWKT=FALSE} can be passed as parameter
}
\examples{
ST_as.WKT(ST_Point(1:2))
}

