% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R, R/sp.R
\name{st_as_sf}
\alias{st_as_sf}
\alias{st_as_sf.Spatial}
\alias{st_as_sf.data.frame}
\title{convert foreign object to an sf object}
\usage{
st_as_sf(x, ...)

\method{st_as_sf}{data.frame}(x, ..., relation_to_geometry = NA_character_,
  coords, third = "XYZ", epsg = NA_integer_, proj4string = NA_character_,
  remove_coordinates = TRUE)

\method{st_as_sf}{Spatial}(x, ...)
}
\arguments{
\item{x}{object to be converted into an object class \code{sf}}

\item{...}{further arguments}

\item{relation_to_geometry}{character vector; see details section of \link{st_sf}}

\item{coords}{in case of point data: coordinate names or numbers}

\item{third}{passed on to \link{st_point} (only when coords is given)}

\item{epsg}{integer denoting the epsg ID (only when coords is given)}

\item{proj4string}{character; giving the proj4 string (only when coords is given)}

\item{remove_coordinates}{logical; when coords is given, remove coordinate columns from data.frame?}
}
\description{
convert foreign object to an sf object
}
\examples{
pt1 = st_point(c(0,1))
pt2 = st_point(c(1,1))
st_sfc(pt1, pt2)
d = data.frame(a = 1:2)
d$geom = st_sfc(pt1, pt2)
df = st_as_sf(d)
d$geom2 = st_sfc(pt1, pt2)
st_as_sf(d) # should warn
data(meuse, package = "sp")
meuse_sf = st_as_sf(meuse, coords = c("x", "y"), epsg = 28992)
meuse_sf[1:3,]
summary(meuse_sf)
library(sp)
x = rbind(c(-1,-1), c(1,-1), c(1,1), c(-1,1), c(-1,-1))
x1 = 0.1 * x + 0.1
x2 = 0.1 * x + 0.4
x3 = 0.1 * x + 0.7
y = x + 3
y1 = x1 + 3
y2 = x2 + 3
y3 = x3 + 3
p = Polygons(list( Polygon(x[5:1,]), Polygon(x2), Polygon(y2), Polygon(x3), 
   Polygon(y[5:1,]), Polygon(y1), Polygon(x1), Polygon(y3)), "ID1")
if (require("rgeos")) {
  r = createSPComment(SpatialPolygons(list(p)))
  comment(r)
  comment(r@polygons[[1]])
  scan(text = comment(r@polygons[[1]]), quiet = TRUE)
  library(sf)
  a = st_as_sf(r)
  summary(a)
}
demo(meuse, ask = FALSE, echo = FALSE)
summary(st_as_sf(meuse))
summary(st_as_sf(meuse.grid))
summary(st_as_sf(meuse.area))
summary(st_as_sf(meuse.riv))
summary(st_as_sf(as(meuse.riv, "SpatialLines")))
pol.grd = as(meuse.grid, "SpatialPolygonsDataFrame")
summary(st_as_sf(pol.grd))
summary(st_as_sf(as(pol.grd, "SpatialLinesDataFrame")))
}

