% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R, R/wkb.R
\name{st_as_sfc}
\alias{st_as_sfc}
\alias{st_as_sfc.SpatialLines}
\alias{st_as_sfc.SpatialMultiPoints}
\alias{st_as_sfc.SpatialPixels}
\alias{st_as_sfc.SpatialPoints}
\alias{st_as_sfc.SpatialPolygons}
\alias{st_as_sfc.WKB}
\title{convert foreign geometry object to an sfc object}
\usage{
st_as_sfc(x, ...)

\method{st_as_sfc}{SpatialPoints}(x, ...)

\method{st_as_sfc}{SpatialPixels}(x, ...)

\method{st_as_sfc}{SpatialMultiPoints}(x, ...)

\method{st_as_sfc}{SpatialLines}(x, ..., forceMulti = FALSE)

\method{st_as_sfc}{SpatialPolygons}(x, ..., forceMulti = FALSE)

\method{st_as_sfc}{WKB}(x, ..., EWKB = FALSE)
}
\arguments{
\item{x}{object to convert}

\item{...}{further arguments}

\item{forceMulti}{logical; if \code{TRUE}, force coercion into \code{MULTIPOLYGON} or \code{MULTILINE} objects, else autodetect}

\item{EWKB}{logical; if TRUE, parse as EWKB (PostGIS: ST_AsEWKB), otherwise as ISO WKB (PostGIS: ST_AsBinary)}
}
\description{
convert foreign geometry object to an sfc object
}

