% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{st_as_text}
\alias{st_as_text}
\alias{st_as_text.sfc}
\alias{st_as_text.sfg}
\title{Return Well-known Text representation of simple feature geometry}
\usage{
st_as_text(x, ...)

\method{st_as_text}{sfg}(x, ...)

\method{st_as_text}{sfc}(x, ..., EWKT = FALSE)
}
\arguments{
\item{x}{object of class sfg or sfc}

\item{...}{passed on to WKT_name}

\item{EWKT}{logical; if TRUE, print SRID=xxx; before the WKT string if epsg is available}
}
\description{
Return Well-known Text representation of simple feature geometry
}
\details{
to suppress printing of SRID, \code{EWKT=FALSE} can be passed as parameter
}
\examples{
st_as_text(st_point(1:2))
}

