% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{st_bbox}
\alias{st_bbox}
\alias{st_bbox.GEOMETRYCOLLECTION}
\alias{st_bbox.LINESTRING}
\alias{st_bbox.MULTILINESTRING}
\alias{st_bbox.MULTIPOINT}
\alias{st_bbox.MULTIPOLYGON}
\alias{st_bbox.POINT}
\alias{st_bbox.POLYGON}
\alias{st_bbox.sf}
\alias{st_bbox.sfc}
\alias{st_bbox.sfc_GEOMETRY}
\alias{st_bbox.sfc_GEOMETRYCOLLECTION}
\alias{st_bbox.sfc_LINESTRING}
\alias{st_bbox.sfc_MULTILINESTRING}
\alias{st_bbox.sfc_MULTIPOINT}
\alias{st_bbox.sfc_MULTIPOLYGON}
\alias{st_bbox.sfc_POINT}
\alias{st_bbox.sfc_POLYGON}
\title{Return bounding of a simple feature or simple feature set}
\usage{
st_bbox(obj)

\method{st_bbox}{POINT}(obj)

\method{st_bbox}{MULTIPOINT}(obj)

\method{st_bbox}{LINESTRING}(obj)

\method{st_bbox}{POLYGON}(obj)

\method{st_bbox}{MULTILINESTRING}(obj)

\method{st_bbox}{MULTIPOLYGON}(obj)

\method{st_bbox}{GEOMETRYCOLLECTION}(obj)

\method{st_bbox}{sfc_POINT}(obj)

\method{st_bbox}{sfc_MULTIPOINT}(obj)

\method{st_bbox}{sfc_LINESTRING}(obj)

\method{st_bbox}{sfc_POLYGON}(obj)

\method{st_bbox}{sfc_MULTILINESTRING}(obj)

\method{st_bbox}{sfc_MULTIPOLYGON}(obj)

\method{st_bbox}{sfc_GEOMETRYCOLLECTION}(obj)

\method{st_bbox}{sfc_GEOMETRY}(obj)

\method{st_bbox}{sfc}(obj)

\method{st_bbox}{sf}(obj)
}
\arguments{
\item{obj}{object to compute the bounding box from}
}
\value{
a numeric vector of length four, with \code{xmin}, \code{ymin}, \code{xmax}
and \code{ymax} values; if \code{obj} is of class \code{sf} or \code{sfc}, the object
returned has a class \code{bbox}, an attribute \code{crs} and a method to print the
bbox and an \code{st_crs} method to retrieve the coordinate reference system 
corresponding to \code{obj} (and hence the bounding box).
}
\description{
Return bounding of a simple feature or simple feature set
}

