% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\docType{data}
\name{st_bbox}
\alias{st_bbox}
\alias{NA_bbox_}
\alias{st_bbox}
\alias{st_bbox}
\alias{st_bbox.POINT}
\alias{st_bbox}
\alias{st_bbox.MULTIPOINT}
\alias{st_bbox}
\alias{st_bbox.LINESTRING}
\alias{st_bbox}
\alias{st_bbox.POLYGON}
\alias{st_bbox}
\alias{st_bbox.MULTILINESTRING}
\alias{st_bbox}
\alias{st_bbox.MULTIPOLYGON}
\alias{st_bbox}
\alias{st_bbox.GEOMETRYCOLLECTION}
\alias{st_bbox}
\alias{st_bbox.MULTISURFACE}
\alias{st_bbox}
\alias{st_bbox.MULTICURVE}
\alias{st_bbox}
\alias{st_bbox.CURVEPOLYGON}
\alias{st_bbox}
\alias{st_bbox.COMPOUNDCURVE}
\alias{st_bbox}
\alias{st_bbox.POLYHEDRALSURFACE}
\alias{st_bbox}
\alias{st_bbox.TIN}
\alias{st_bbox}
\alias{st_bbox.TRIANGLE}
\alias{st_bbox}
\alias{st_bbox.CIRCULARSTRING}
\alias{st_bbox}
\alias{st_bbox.sfc}
\alias{st_bbox}
\alias{st_bbox.sf}
\title{Return bounding of a simple feature or simple feature set}
\format{An object of class \code{bbox} of length 4.}
\usage{
NA_bbox_

st_bbox(obj)

\\method{st_bbox}{POINT}(obj)

\\method{st_bbox}{MULTIPOINT}(obj)

\\method{st_bbox}{LINESTRING}(obj)

\\method{st_bbox}{POLYGON}(obj)

\\method{st_bbox}{MULTILINESTRING}(obj)

\\method{st_bbox}{MULTIPOLYGON}(obj)

\\method{st_bbox}{GEOMETRYCOLLECTION}(obj)

\\method{st_bbox}{MULTISURFACE}(obj)

\\method{st_bbox}{MULTICURVE}(obj)

\\method{st_bbox}{CURVEPOLYGON}(obj)

\\method{st_bbox}{COMPOUNDCURVE}(obj)

\\method{st_bbox}{POLYHEDRALSURFACE}(obj)

\\method{st_bbox}{TIN}(obj)

\\method{st_bbox}{TRIANGLE}(obj)

\\method{st_bbox}{CIRCULARSTRING}(obj)

\\method{st_bbox}{sfc}(obj)

\\method{st_bbox}{sf}(obj)
}
\arguments{
\item{obj}{object to compute the bounding box from}
}
\value{
a numeric vector of length four, with \code{xmin}, \code{ymin}, \code{xmax}
and \code{ymax} values; if \code{obj} is of class \code{sf} or \code{sfc}, the object
returned has a class \code{bbox}, an attribute \code{crs} and a method to print the
bbox and an \code{st_crs} method to retrieve the coordinate reference system 
corresponding to \code{obj} (and hence the bounding box).
}
\description{
Return bounding of a simple feature or simple feature set
}
\details{
\code{NA_bbox_} is the \code{bbox} object with a missing value.
}
\keyword{datasets}
