% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{st_cast}
\alias{st_cast}
\title{Cast geometry to another type: either simplify, or cast explicitly}
\usage{
st_cast(x, to, ...)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfc} or \code{sf}}

\item{to}{character; target type, if missing, simplification is tried; when \code{x} is of type \code{sfg} (i.e., a single geometry) then \code{to} needs to be specified.}

\item{...}{ignored}
}
\value{
object of class \code{to} if successful, or unmodified object if unsuccesful. If information gets lost while type casting, a warning is raised.
}
\description{
Cast geometry to another type: either simplify, or cast explicitly
}
\examples{
s = st_multipoint(rbind(c(1,0)))
st_cast(s, "POINT")
}

