% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{st_coordinates}
\alias{st_coordinates}
\title{retrieve coordinates in matrix form}
\usage{
st_coordinates(x, ...)
}
\arguments{
\item{x}{object of class sf, sfc or sfg}

\item{...}{ignored}
}
\description{
retrieve coordinates in matrix form
}
