% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{st_drivers}
\alias{st_drivers}
\title{Get GDAL drivers}
\usage{
st_drivers(what = "vector")
}
\arguments{
\item{what}{character: "vector" or "raster", anything else will return all drivers.}
}
\value{
a \code{data.frame} with driver metadata
}
\description{
Get a list of the available GDAL drivers
}
\details{
The drivers available will depend on the installation of GDAL/OGR, and can vary; the \code{st_drivers()}
function shows which are available, and which may be written (but all are assumed to be readable). Note that stray
files in data source directories (such as *.dbf) may lead to suprious errors that accompanying *.shp are missing.
}
\examples{
st_drivers()
}
