% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{st_is_longlat}
\alias{st_is_longlat}
\title{assert whether simple feature coordinates are longlat degrees}
\usage{
st_is_longlat(x)
}
\arguments{
\item{x}{object of class \link{sf} or \link{sfc}}
}
\value{
TRUE if \code{+proj=longlat} is part of the proj4string, NA if this string is missing, FALSE otherwise
}
\description{
assert whether simple feature coordinates are longlat degrees
}

