% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{st_is_longlat}
\alias{st_is_longlat}
\title{Assert whether simple feature coordinates are longlat degrees}
\usage{
st_is_longlat(x)
}
\arguments{
\item{x}{object of class \link{sf} or \link{sfc}, or otherwise an object of a class that has an \link{st_crs} method returning a \code{crs} object}
}
\value{
\code{TRUE} if \code{x} has geographic coordinates, \code{FALSE} if it has projected coordinates, or \code{NA} if \code{is.na(st_crs(x))}.
}
\description{
Assert whether simple feature coordinates are longlat degrees
}
