% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{st_precision}
\alias{st_precision}
\alias{st_precision<-}
\alias{st_set_precision}
\title{Get precision}
\usage{
st_precision(x)

st_set_precision(x, precision)

st_precision(x) <- value
}
\arguments{
\item{x}{sfc}

\item{precision}{numeric. See details}

\item{x}{sfc}
}
\description{
Get precision

Set precision
}
\details{
Precision is a decimal number that sets the number of decimals to 
use in geometric operations. Default uses all decimals (0). To specify 3 
decimals, use 0.0001. It applies to all dimensions (x, y, z, m). Setting 
precision has no impact on the stored 
values.
}
\examples{
x <- st_sfc(st_point(c(pi, pi)))
st_precision(x)
st_precision(x) <- 0.01
st_precision(x)
}

