% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{st_read}
\alias{st_read}
\alias{st_read_db}
\title{read simple features from file or database}
\usage{
st_read(dsn, layer, ..., options = NULL, quiet = FALSE, iGeomField = 1L,
  type = 0)

st_read_db(conn = NULL, table, query = paste("select * from ", table, ";"),
  geom_column = NULL, ...)
}
\arguments{
\item{dsn}{data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder)}

\item{layer}{layer name (varies by driver, may be a file name without extension)}

\item{...}{parameter(s) passed on to \link{st_as_sf}}

\item{options}{character; driver dependent layer reading options; multiple options supported.}

\item{quiet}{logical; suppress info on name, driver, size and spatial reference}

\item{iGeomField}{integer; in case of multiple geometry fields, which one to take?}

\item{type}{integer; ISO number of desired simple feature type; see details. If left zero, in case of mixed feature geometry types, conversion to the highest numeric type value found will be attempted.}

\item{conn}{open database connection}

\item{table}{table name}

\item{query}{SQL query to select records}

\item{geom_column}{character or integer: indicator of name or position of the geometry column; if not provided, the last column of type character is chosen}
}
\value{
object of class \link{sf}
}
\description{
read simple features from file or database

read PostGIS table directly through DBI and RPostgreSQL interface, converting binary
}
\details{
for iGeomField, see also \url{https://trac.osgeo.org/gdal/wiki/rfc41_multiple_geometry_fields}; for \code{type} values see \url{https://en.wikipedia.org/wiki/Well-known_text#Well-known_binary}, but note that not every target value may lead to succesful conversion. The typical conversion from POLYGON (3) to MULTIPOLYGON (6) should work; the other way around (type=3), secondary rings from MULTIPOLYGONS may be dropped without warnings.
}
\examples{
if (Sys.getenv("USER") \%in\% c("edzer", "travis")) { # load meuse to postgis
 library(sp)
 example(meuse, ask = FALSE, echo = FALSE)
 st_write(st_as_sf(meuse), "PG:dbname=postgis", "meuse", driver = "PostgreSQL", 
   options = "OVERWRITE=true")
 (s = st_read("PG:dbname=postgis", "meuse"))
 summary(s)
}
nc = st_read(system.file("gpkg/nc.gpkg", package="sf"), "nc.gpkg", crs = 4267)
summary(nc)
if (Sys.getenv("USER") \%in\% c("travis", "edzer")) {
  library(RPostgreSQL)
  conn = dbConnect(PostgreSQL(), dbname = "postgis")
  st_read_db(conn, query = "select * from meuse limit 3;")
  dbDisconnect(conn)
}
}

