% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{st_read}
\alias{st_read}
\alias{st_read_db}
\title{read simple features from file or database}
\usage{
st_read(dsn, layer, ..., quiet = FALSE, iGeomField = 1L)

st_read_db(conn = NULL, table, query = paste("select * from ", table, ";"),
  geom_column = NULL, ...)
}
\arguments{
\item{dsn}{data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder)}

\item{layer}{layer name (varies by driver, may be a file name without extension)}

\item{...}{parameter(s) passed on to \link{st_as_sf}}

\item{quiet}{logical; suppress info on name, driver, size and spatial reference}

\item{iGeomField}{integer; in case of multiple geometry fields, which one to take?}

\item{conn}{open database connection}

\item{table}{table name}

\item{query}{SQL query to select records}

\item{geom_column}{character or integer: indicator of name or position of the geometry column; if not provided, the last column of type character is chosen}
}
\description{
read simple features from file or database

read PostGIS table directly through DBI and RPostgreSQL interface, converting wkb
}
\details{
for iGeomField, see also \url{https://trac.osgeo.org/gdal/wiki/rfc41_multiple_geometry_fields}
}
\examples{
if (Sys.getenv("USER") \%in\% c("edzer", "travis")) { # load meuse to postgis
 library(sp)
 example(meuse, ask = FALSE, echo = FALSE)
 st_write(st_as_sf(meuse), "PG:dbname=postgis", "meuse", driver = "PostgreSQL")
 (s = st_read("PG:dbname=postgis", "meuse"))
 summary(s)
}
(s = st_read(system.file("shapes/", package="maptools"), "sids"))[1:10,]
summary(s)
if (Sys.getenv("USER") \%in\% c("travis", "edzer")) {
  library(RPostgreSQL)
  conn = dbConnect(PostgreSQL(), dbname = "postgis")
  st_read_db(conn, query = "select * from meuse limit 3;")
  dbDisconnect(conn)
}
}

