% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ST_read}
\alias{ST_read}
\alias{ST_write}
\title{read simple features from file or database}
\usage{
ST_read(dsn, layer, ...)

ST_write(sf, dsn = ".", layer, driver = "ESRI Shapefile",
  opts = character(), ...)
}
\arguments{
\item{dsn}{data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder)}

\item{layer}{layer name (varies by driver, may be a file name without extension)}

\item{...}{parameter(s) passed on to \link{ST_as.sf}}

\item{sf}{object of class \code{sf}}

\item{driver}{driver name}

\item{opts}{options to pass on to driver}
}
\description{
read simple features from file or database
}
\examples{
if (Sys.getenv("USER") == "edzer") {
 (s = ST_read("PG:dbname=test", "test"))
 summary(s)
}
(s = ST_read(system.file("shapes/", package="maptools"), "sids"))[1:10,]
summary(s)
}

