% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{stars}
\alias{stars}
\alias{.get_layout}
\alias{stars}
\alias{.degAxis}
\alias{stars}
\alias{.image_scale}
\alias{stars}
\alias{.image_scale_factor}
\title{functions only exported to be used internally by stars}
\usage{
.get_layout(bb, n, total_size, key.pos, key.size)

.degAxis(side, at, labels, ..., lon, lat, ndiscr, reset)

.image_scale(z, col, breaks = NULL, key.pos, add.axis = TRUE, at = NULL,
  ..., axes = FALSE)

.image_scale_factor(z, col, breaks = NULL, key.pos, add.axis = TRUE,
  at = NULL, ..., axes = FALSE, key.size)
}
\arguments{
\item{bb}{ignore}

\item{n}{ignore}

\item{total_size}{ignore}

\item{key.pos}{ignore}

\item{key.size}{ignore}

\item{side}{ignore}

\item{at}{ignore}

\item{labels}{ignore}

\item{...}{ignore}

\item{lon}{ignore}

\item{lat}{ignore}

\item{ndiscr}{ignore}

\item{reset}{ignore}

\item{z}{ignore}

\item{col}{ignore}

\item{breaks}{ignore}

\item{add.axis}{ignore}

\item{axes}{ignore}
}
\description{
functions only exported to be used internally by stars
}
