% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{tibble}
\alias{tibble}
\alias{type_sum.sfc}
\alias{tibble}
\alias{obj_sum.sfc}
\title{Summarize simple feature type for tibble}
\usage{
type_sum.sfc(x, ...)

obj_sum.sfc(x)
}
\arguments{
\item{x}{object of class sfc}

\item{...}{ignored}
}
\description{
Summarize simple feature type for tibble

Summarize simple feature item for tibble
}
